@@ -18,11 +18,14 @@ import (
 	"github.com/buger/jsonparser"
 	"github.com/hashicorp/go-retryablehttp"
 	"github.com/mattbaird/jsonpatch"
+	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
+	otrace "go.opentelemetry.io/otel/trace"
 	"go.uber.org/zap"
 
 	"github.com/wundergraph/wundergraph/pkg/authentication"
 	"github.com/wundergraph/wundergraph/pkg/logging"
 	"github.com/wundergraph/wundergraph/pkg/pool"
+	"github.com/wundergraph/wundergraph/pkg/trace"
 )
 
 type WunderGraphRequest struct {
@@ -160,17 +163,34 @@ type Client struct {
 	log                 *zap.Logger
 }
 
-func NewClient(serverUrl string, logger *zap.Logger) *Client {
+type ClientOptions struct {
+	ServerURL     string
+	EnableTracing bool
+	Logger        *zap.Logger
+}
+
+func NewClient(opts *ClientOptions) *Client {
+
+	rt := http.DefaultTransport
+
+	if opts.EnableTracing {
+		rt = trace.NewTransport(rt,
+			otelhttp.WithSpanOptions(otrace.WithAttributes(trace.HooksClientAttribute)),
+			otelhttp.WithSpanNameFormatter(trace.SpanNameFormatter),
+		)
+	}
+
 	return &Client{
-		serverUrl:           serverUrl,
-		httpClient:          buildClient(60 * time.Second),
-		subscriptionsClient: buildClient(0),
-		log:                 logger,
+		serverUrl:           opts.ServerURL,
+		httpClient:          buildClient(60*time.Second, rt),
+		subscriptionsClient: buildClient(0, rt),
+		log:                 opts.Logger,
 	}
 }
 
-func buildClient(requestTimeout time.Duration) *retryablehttp.Client {
+func buildClient(requestTimeout time.Duration, rt http.RoundTripper) *retryablehttp.Client {
 	httpClient := retryablehttp.NewClient()
+	httpClient.HTTPClient.Transport = rt
 	// we will try 40 times with a constant delay of 50ms after max 2s we will give up
 	httpClient.RetryMax = 40
 	// keep it low and linear to increase the chance
@@ -179,7 +199,7 @@ func buildClient(requestTimeout time.Duration) *retryablehttp.Client {
 	httpClient.RetryWaitMax = 50 * time.Millisecond
 	httpClient.RetryWaitMin = 50 * time.Millisecond
 	httpClient.HTTPClient.Timeout = requestTimeout
-	httpClient.Logger = log.New(ioutil.Discard, "", log.LstdFlags)
+	httpClient.Logger = log.New(io.Discard, "", log.LstdFlags)
 	httpClient.CheckRetry = func(ctx context.Context, resp *http.Response, err error) (bool, error) {
 		if resp != nil && resp.StatusCode == http.StatusInternalServerError {
 			return false, nil
