@@ -33,7 +33,12 @@ import { OpenApiServerConfig } from './plugins/omnigraphOAS';
 import { SoapServerConfig } from './plugins/omnigraphSOAP';
 import { NamespacingExecutor } from '../orm';
 import type { OperationsAsyncContext } from './operations-context';
+import configureTracerProvider from './trace/trace';
+import { propagation } from '@opentelemetry/api';
+import { TracerConfig } from './types';
+import { loadTraceConfigFromWgConfig } from './trace/config';
 
+const isProduction = process.env.NODE_ENV === 'production';
 let WG_CONFIG: WunderGraphConfiguration;
 let clientFactory: InternalClientFactory;
 let logger: pino.Logger;
@@ -129,8 +134,6 @@ const _configureWunderGraphServer = <
 	 * This environment variable is used to determine if the server should start the hooks server.
 	 */
 	if (process.env.START_HOOKS_SERVER === 'true') {
-		const isProduction = process.env.NODE_ENV === 'production';
-
 		if (!isProduction) {
 			// Exit the server when wunderctl exited without the chance to kill the child processes
 			onParentProcessExit(() => {
@@ -163,6 +166,7 @@ export const startServer = async (opts: ServerRunOptions) => {
 		const port = parseInt(portString, 10);
 
 		const fastify = await createServer(opts);
+
 		await fastify.listen({
 			port: port,
 			host: host,
@@ -289,6 +293,41 @@ export const createServer = async ({
 	});
 
 	/**
+	 * Enable telemetry
+	 */
+
+	if (config.api?.nodeOptions?.openTelemetry) {
+		const tracerConfig: TracerConfig = loadTraceConfigFromWgConfig(config.api?.nodeOptions.openTelemetry);
+
+		if (tracerConfig.enabled) {
+			let batchTimeoutMs = isProduction ? 3000 : 1000;
+			const batchTimeoutMsEnv = process.env[WgEnv.OtelBatchTimeoutMs];
+			if (batchTimeoutMsEnv) {
+				batchTimeoutMs = parseInt(batchTimeoutMsEnv);
+			}
+
+			fastify.log.info(
+				{
+					endpoint: tracerConfig.httpEndpoint,
+					batchTimeoutMs,
+				},
+				'OpenTelemetry enabled'
+			);
+
+			const tracerProvider = configureTracerProvider(
+				{
+					httpEndpoint: tracerConfig.httpEndpoint,
+					authToken: tracerConfig.authToken,
+					batchTimeoutMs,
+				},
+				logger
+			);
+
+			await fastify.register(require('./plugins/telemetry'), tracerProvider);
+		}
+	}
+
+	/**
 	 * We encapsulate the preHandler with a fastify plugin "register" to not apply it on other routes.
 	 */
 	await fastify.register(async (fastify) => {
@@ -299,17 +338,25 @@ export const createServer = async ({
 		fastify.addHook<{ Body: FastifyRequestBody }>('preHandler', async (req, reply) => {
 			// clientRequest represents the original client request that was sent initially to the WunderNode.
 			const clientRequest = createClientRequest(req.body);
+
+			const headers: { [key: string]: string } = {
+				'x-request-id': req.id,
+			};
+			if (req.telemetry) {
+				propagation.inject(req.telemetry.context, headers);
+			}
+
 			req.ctx = {
 				log: req.log,
 				user: req.body.__wg.user!,
 				clientRequest,
-				internalClient: clientFactory({ 'x-request-id': req.id }, req.body.__wg.clientRequest),
+				internalClient: clientFactory(headers, req.body.__wg.clientRequest),
 				operations: new OperationsClient({
 					baseURL: nodeInternalURL,
 					clientRequest: req.body.__wg.clientRequest,
-					extraHeaders: {
-						'x-request-id': req.id,
-					},
+					extraHeaders: headers,
+					tracer: fastify.tracer,
+					traceContext: req.telemetry?.context,
 				}),
 				context: await createContext(globalContext),
 			};
