@@ -2,17 +2,17 @@ import { defineWorkspace } from 'vitest/config';
 
 export default defineWorkspace([
 	{
-		root: 'apps/basic',
+		root: 'apps/auth',
 		test: {
-			globalSetup: 'apps/basic/test-setup.ts',
+			globalSetup: 'apps/auth/test-setup.ts',
 			testTimeout: 15000,
 			hookTimeout: 20000,
 		},
 	},
 	{
-		root: 'apps/auth',
+		root: 'apps/basic',
 		test: {
-			globalSetup: 'apps/auth/test-setup.ts',
+			globalSetup: 'apps/basic/test-setup.ts',
 			testTimeout: 15000,
 			hookTimeout: 20000,
 		},
@@ -29,7 +29,8 @@ export default defineWorkspace([
 		root: 'apps/http-proxy',
 		test: {
 			globalSetup: 'apps/http-proxy/test-setup.ts',
-			testTimeout: 15000,
+			// High timeout because we're testing code generation
+			testTimeout: 60000,
 			hookTimeout: 20000,
 		},
 	},
@@ -45,6 +46,14 @@ export default defineWorkspace([
 		root: 'apps/openapi',
 		test: {
 			globalSetup: 'apps/openapi/test-setup.ts',
+			testTimeout: 15000,
+			hookTimeout: 20000,
+		},
+	},
+	{
+		root: 'apps/opentelemetry',
+		test: {
+			globalSetup: 'apps/opentelemetry/test-setup.ts',
 			testTimeout: 15000,
 			hookTimeout: 20000,
 		},
