@@ -47,9 +47,11 @@ import (
 	"github.com/wundergraph/wundergraph/pkg/loadvariable"
 	"github.com/wundergraph/wundergraph/pkg/logging"
 	"github.com/wundergraph/wundergraph/pkg/metrics"
+	"github.com/wundergraph/wundergraph/pkg/operation"
 	"github.com/wundergraph/wundergraph/pkg/pool"
 	"github.com/wundergraph/wundergraph/pkg/postresolvetransform"
 	"github.com/wundergraph/wundergraph/pkg/s3uploadclient"
+	"github.com/wundergraph/wundergraph/pkg/trace"
 	"github.com/wundergraph/wundergraph/pkg/webhookhandler"
 	"github.com/wundergraph/wundergraph/pkg/wgpb"
 )
@@ -149,6 +151,10 @@ func (r *Builder) BuildAndMountApiHandler(ctx context.Context, router *mux.Route
 
 	r.router = r.createSubRouter(router)
 
+	if r.enableRequestLogging {
+		r.router.Use(logRequestMiddleware(os.Stderr))
+	}
+
 	for _, webhook := range api.Webhooks {
 		err = r.registerWebhook(webhook)
 		if err != nil {
@@ -194,43 +200,9 @@ func (r *Builder) BuildAndMountApiHandler(ctx context.Context, router *mux.Route
 		r.planConfig.DataSources = append(r.planConfig.DataSources, dataSource)
 	}
 
-	// limiter := rate.NewLimiter(rate.Every(time.Second), 10)
-
 	r.log.Debug("configuring API",
 		zap.Int("numOfOperations", len(api.Operations)),
 	)
-
-	// Redirect from old-style URL, for temporary backwards compatibility
-	r.router.MatcherFunc(func(r *http.Request, rm *mux.RouteMatch) bool {
-		components := strings.Split(r.URL.Path, "/")
-		return len(components) > 2 && components[2] == "main"
-	}).HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
-		components := strings.Split(req.URL.Path, "/")
-		prefix := strings.Join(components[:3], "/")
-		const format = "URLs with the %q prefix are deprecated and will be removed in a future release, " +
-			"see https://github.com/wundergraph/wundergraph/blob/main/docs/migrations/sdk-0.122.0-0.123.0.md"
-		r.log.Warn(fmt.Sprintf(format, prefix), zap.String("URL", req.URL.Path))
-		req.URL.Path = "/" + strings.Join(components[3:], "/")
-		r.router.ServeHTTP(w, req)
-	})
-
-	if len(api.Hosts) > 0 {
-		r.router.Use(func(handler http.Handler) http.Handler {
-			return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-				for i := range api.Hosts {
-					if r.Host == api.Hosts[i] {
-						handler.ServeHTTP(w, r)
-						return
-					}
-				}
-				http.Error(w, fmt.Sprintf("Host not found: %s", r.Host), http.StatusNotFound)
-			})
-		})
-	}
-
-	if r.enableRequestLogging {
-		r.router.Use(logRequestMiddleware(os.Stderr))
-	}
 
 	r.router.Use(func(handler http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, request *http.Request) {
@@ -241,6 +213,18 @@ func (r *Builder) BuildAndMountApiHandler(ctx context.Context, router *mux.Route
 			}
 
 			request = request.WithContext(context.WithValue(request.Context(), logging.RequestIDKey{}, requestID))
+
+			if len(api.Hosts) > 0 {
+				for i := range api.Hosts {
+					if request.Host == api.Hosts[i] {
+						handler.ServeHTTP(w, request)
+						return
+					}
+				}
+				http.Error(w, fmt.Sprintf("Host not found: %s", request.Host), http.StatusNotFound)
+				return
+			}
+
 			handler.ServeHTTP(w, request)
 		})
 	})
@@ -924,7 +908,10 @@ type QueryHandler struct {
 
 func (h *QueryHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	requestLogger := h.log.With(logging.WithRequestIDFromContext(r.Context()))
-	r = setOperationMetaData(r, h.operation)
+	r = operation.SetOperationMetaData(r, h.operation)
+
+	// Set trace attributes based on the current operation
+	trace.SetOperationAttributes(r.Context())
 
 	if proceed := h.rbacEnforcer.Enforce(r); !proceed {
 		http.Error(w, "unauthorized", http.StatusUnauthorized)
@@ -936,36 +923,36 @@ func (h *QueryHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	buf := pool.GetBytesBuffer()
 	defer pool.PutBytesBuffer(buf)
 
-	ctx := pool.GetCtx(r, r, pool.Config{
+	resolveCtx := pool.GetCtx(r, r, pool.Config{
 		RenameTypeNames: h.renameTypeNames,
 	})
-	defer pool.PutCtx(ctx)
+	defer pool.PutCtx(resolveCtx)
 
-	ctx.Variables = parseQueryVariables(r, h.queryParamsAllowList)
-	ctx.Variables = h.stringInterpolator.Interpolate(ctx.Variables)
+	resolveCtx.Variables = parseQueryVariables(r, h.queryParamsAllowList)
+	resolveCtx.Variables = h.stringInterpolator.Interpolate(resolveCtx.Variables)
 
-	if !validateInputVariables(ctx.Context(), requestLogger, ctx.Variables, h.variablesValidator, w) {
+	if !validateInputVariables(resolveCtx.Context(), requestLogger, resolveCtx.Variables, h.variablesValidator, w) {
 		return
 	}
 
 	compactBuf := pool.GetBytesBuffer()
 	defer pool.PutBytesBuffer(compactBuf)
-	err := json.Compact(compactBuf, ctx.Variables)
+	err := json.Compact(compactBuf, resolveCtx.Variables)
 	if err != nil {
 		requestLogger.Error("Could not compact variables in query handler", zap.Bool("isLive", isLive), zap.Error(err))
 		w.WriteHeader(http.StatusBadRequest)
 		return
 	}
-	ctx.Variables = compactBuf.Bytes()
+	resolveCtx.Variables = compactBuf.Bytes()
 
-	ctx.Variables = h.jsonStringInterpolator.Interpolate(ctx.Variables)
+	resolveCtx.Variables = h.jsonStringInterpolator.Interpolate(resolveCtx.Variables)
 
 	if len(h.extractedVariables) != 0 {
 		// we make a copy of the extracted variables to not override h.extractedVariables
-		ctx.Variables = MergeJsonRightIntoLeft(h.extractedVariables, ctx.Variables)
+		resolveCtx.Variables = MergeJsonRightIntoLeft(h.extractedVariables, resolveCtx.Variables)
 	}
 
-	ctx.Variables, err = postProcessVariables(h.operation, r, ctx.Variables)
+	resolveCtx.Variables, err = postProcessVariables(h.operation, r, resolveCtx.Variables)
 	if err != nil {
 		if done := handleOperationErr(requestLogger, err, w, "postProcessVariables failed", h.operation); done {
 			return
@@ -986,11 +973,11 @@ func (h *QueryHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	}
 
 	if isLive {
-		h.handleLiveQuery(r, w, ctx, buf, flusher, requestLogger)
+		h.handleLiveQuery(r, w, resolveCtx, buf, flusher, requestLogger)
 		return
 	}
 
-	resp, err := h.hooksPipeline.Run(ctx, w, r, buf)
+	resp, err := h.hooksPipeline.Run(resolveCtx, w, r, buf)
 	if done := handleOperationErr(requestLogger, err, w, "hooks pipeline failed", h.operation); done {
 		return
 	}
@@ -1188,7 +1175,10 @@ func (h *MutationHandler) parseFormVariables(r *http.Request) []byte {
 
 func (h *MutationHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	requestLogger := h.log.With(logging.WithRequestIDFromContext(r.Context()))
-	r = setOperationMetaData(r, h.operation)
+	r = operation.SetOperationMetaData(r, h.operation)
+
+	// Set trace attributes based on the current operation
+	trace.SetOperationAttributes(r.Context())
 
 	if proceed := h.rbacEnforcer.Enforce(r); !proceed {
 		http.Error(w, "unauthorized", http.StatusUnauthorized)
@@ -1295,7 +1285,10 @@ type SubscriptionHandler struct {
 
 func (h *SubscriptionHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	requestLogger := h.log.With(logging.WithRequestIDFromContext(r.Context()))
-	r = setOperationMetaData(r, h.operation)
+	r = operation.SetOperationMetaData(r, h.operation)
+
+	// Set trace attributes based on the current operation
+	trace.SetOperationAttributes(r.Context())
 
 	// recover from panic if one occured. Set err to nil otherwise.
 	defer func() {
@@ -1792,7 +1785,10 @@ func (h *FunctionsHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	requestLogger := h.log.With(logging.WithRequestID(reqID))
 	r = r.WithContext(context.WithValue(r.Context(), logging.RequestIDKey{}, reqID))
 
-	r = setOperationMetaData(r, h.operation)
+	r = operation.SetOperationMetaData(r, h.operation)
+
+	// Set trace attributes based on the current operation
+	trace.SetOperationAttributes(r.Context())
 
 	if proceed := h.rbacEnforcer.Enforce(r); !proceed {
 		http.Error(w, "unauthorized", http.StatusUnauthorized)
@@ -2016,47 +2012,6 @@ type EndpointUnavailableHandler struct {
 func (m *EndpointUnavailableHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	m.Logger.Error("operation not available", zap.String("operationName", m.OperationName), zap.String("URL", r.URL.Path))
 	http.Error(w, fmt.Sprintf("Endpoint not available for Operation: %s, please check the logs.", m.OperationName), http.StatusServiceUnavailable)
-}
-
-type OperationMetaData struct {
-	OperationName string
-	OperationType wgpb.OperationType
-}
-
-func (o *OperationMetaData) GetOperationTypeString() string {
-	switch o.OperationType {
-	case wgpb.OperationType_MUTATION:
-		return "mutation"
-	case wgpb.OperationType_QUERY:
-		return "query"
-	case wgpb.OperationType_SUBSCRIPTION:
-		return "subscription"
-	default:
-		return "unknown"
-	}
-}
-
-func setOperationMetaData(r *http.Request, operation *wgpb.Operation) *http.Request {
-	metaData := &OperationMetaData{
-		OperationName: operation.Name,
-		OperationType: operation.OperationType,
-	}
-	return r.WithContext(context.WithValue(r.Context(), "operationMetaData", metaData))
-}
-
-func getOperationMetaData(r *http.Request) *OperationMetaData {
-	if r == nil {
-		return nil
-	}
-	ctx := r.Context()
-	if ctx == nil {
-		return nil
-	}
-	maybeMetaData := r.Context().Value("operationMetaData")
-	if maybeMetaData == nil {
-		return nil
-	}
-	return maybeMetaData.(*OperationMetaData)
 }
 
 func setSubscriptionHeaders(w http.ResponseWriter) {
