@@ -14,13 +14,17 @@ import (
 	"github.com/buger/jsonparser"
 	"github.com/golang-jwt/jwt/v4"
 	"github.com/gorilla/websocket"
+	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
+	otrace "go.opentelemetry.io/otel/trace"
 
 	"github.com/wundergraph/wundergraph/pkg/authentication"
 	"github.com/wundergraph/wundergraph/pkg/hooks"
 	"github.com/wundergraph/wundergraph/pkg/loadvariable"
 	"github.com/wundergraph/wundergraph/pkg/logging"
 	"github.com/wundergraph/wundergraph/pkg/metrics"
+	"github.com/wundergraph/wundergraph/pkg/operation"
 	"github.com/wundergraph/wundergraph/pkg/pool"
+	"github.com/wundergraph/wundergraph/pkg/trace"
 	"github.com/wundergraph/wundergraph/pkg/wgpb"
 )
 
@@ -35,7 +39,17 @@ type apiTransportFactory struct {
 }
 
 func (f *apiTransportFactory) RoundTripper(transport *http.Transport, enableStreamingMode bool) http.RoundTripper {
-	return NewApiTransport(transport, enableStreamingMode, f.opts)
+	rt := NewApiTransport(transport, enableStreamingMode, f.opts)
+
+	if f.opts.EnableTracing {
+		return trace.NewTransport(
+			rt,
+			otelhttp.WithSpanOptions(otrace.WithAttributes(trace.ApiTransportAttribute)),
+			otelhttp.WithSpanNameFormatter(trace.SpanNameFormatter),
+		)
+	}
+
+	return rt
 }
 
 func (f *apiTransportFactory) DefaultTransportTimeout() time.Duration {
@@ -68,6 +82,7 @@ type ApiTransportOptions struct {
 	API                  *Api
 	HooksClient          *hooks.Client
 	EnableRequestLogging bool
+	EnableTracing        bool
 	Metrics              metrics.Metrics
 }
 
@@ -156,7 +171,7 @@ func (t *ApiTransport) roundTrip(request *http.Request, buf *bytes.Buffer) (res 
 	// if you're doing this after calling the onRequest hook, it won't work
 	isUpgradeRequest := websocket.IsWebSocketUpgrade(request)
 
-	metaData := getOperationMetaData(request)
+	metaData := operation.GetOperationMetaData(request.Context())
 	if metaData != nil {
 		_, onRequestHook = t.onRequestHook[metaData.OperationName]
 		_, onResponseHook = t.onResponseHook[metaData.OperationName]
@@ -280,7 +295,7 @@ func (t *ApiTransport) internalGraphQLRoundTrip(request *http.Request) (res *htt
 	return t.httpTransport.RoundTrip(req)
 }
 
-func (t *ApiTransport) handleOnRequestHook(r *http.Request, metaData *OperationMetaData, buf *bytes.Buffer) (*http.Request, error) {
+func (t *ApiTransport) handleOnRequestHook(r *http.Request, metaData *operation.OperationMetaData, buf *bytes.Buffer) (*http.Request, error) {
 	var (
 		body []byte
 		err  error
@@ -339,7 +354,7 @@ func (t *ApiTransport) handleOnRequestHook(r *http.Request, metaData *OperationM
 	return r, nil
 }
 
-func (t *ApiTransport) handleOnResponseHook(r *http.Response, metaData *OperationMetaData, buf *bytes.Buffer) (*http.Response, error) {
+func (t *ApiTransport) handleOnResponseHook(r *http.Response, metaData *operation.OperationMetaData, buf *bytes.Buffer) (*http.Response, error) {
 	body, err := io.ReadAll(r.Body)
 	if err != nil {
 		return nil, err
