@@ -7,17 +7,10 @@ import (
 	"strings"
 	"time"
 
-	"github.com/docker/go-units"
-
 	"github.com/wundergraph/wundergraph/pkg/apihandler"
 	"github.com/wundergraph/wundergraph/pkg/loadvariable"
 	"github.com/wundergraph/wundergraph/pkg/logging"
 	"github.com/wundergraph/wundergraph/pkg/wgpb"
-)
-
-const (
-	defaultInMemoryCacheSize    = int64(128 * units.MB)
-	wgInMemoryCacheConfigEnvKey = "WG_IN_MEMORY_CACHE"
 )
 
 type Server struct {
@@ -99,6 +92,15 @@ func CreateConfig(graphConfig *wgpb.WunderGraphConfiguration) (*WunderNodeConfig
 		return nil, err
 	}
 
+	otelEnabled, err := loadvariable.Bool(graphConfig.Api.NodeOptions.OpenTelemetry.Enabled)
+	if err != nil {
+		return nil, err
+	}
+	otelSampler, err := loadvariable.Float64(graphConfig.Api.NodeOptions.OpenTelemetry.Sampler)
+	if err != nil {
+		return nil, err
+	}
+
 	config := WunderNodeConfig{
 		Api: &apihandler.Api{
 			PrimaryHost:           fmt.Sprintf("%s:%d", listener.Host, listener.Port),
@@ -126,6 +128,12 @@ func CreateConfig(graphConfig *wgpb.WunderGraphConfiguration) (*WunderNodeConfig
 				Prometheus: apihandler.PrometheusOptions{
 					Enabled: prometheusEnabled,
 					Port:    prometheusPort,
+				},
+				OpenTelemetry: apihandler.OpenTelemetry{
+					Enabled:              otelEnabled,
+					AuthToken:            loadvariable.String(graphConfig.Api.NodeOptions.OpenTelemetry.AuthToken),
+					ExporterHTTPEndpoint: loadvariable.String(graphConfig.Api.NodeOptions.OpenTelemetry.ExporterHttpEndpoint),
+					Sampler:              otelSampler,
 				},
 			},
 		},
