@@ -4880,7 +4880,8 @@ type NodeOptions struct {
 	ListenInternal               *InternalListenerOptions `protobuf:"bytes,6,opt,name=listenInternal,proto3" json:"listenInternal,omitempty"`
 	NodeInternalUrl              *ConfigurationVariable   `protobuf:"bytes,7,opt,name=nodeInternalUrl,proto3" json:"nodeInternalUrl,omitempty"`
 	DefaultHttpProxyUrl          *ConfigurationVariable   `protobuf:"bytes,8,opt,name=defaultHttpProxyUrl,proto3" json:"defaultHttpProxyUrl,omitempty"`
-	Prometheus                   *PrometheusOptions       `protobuf:"bytes,9,opt,name=prometheus,proto3" json:"prometheus,omitempty"`
+	OpenTelemetry                *TelemetryOptions        `protobuf:"bytes,9,opt,name=openTelemetry,proto3" json:"openTelemetry,omitempty"`
+	Prometheus                   *PrometheusOptions       `protobuf:"bytes,10,opt,name=prometheus,proto3" json:"prometheus,omitempty"`
 }
 
 func (x *NodeOptions) Reset() {
@@ -4971,9 +4972,87 @@ func (x *NodeOptions) GetDefaultHttpProxyUrl() *ConfigurationVariable {
 	return nil
 }
 
+func (x *NodeOptions) GetOpenTelemetry() *TelemetryOptions {
+	if x != nil {
+		return x.OpenTelemetry
+	}
+	return nil
+}
+
 func (x *NodeOptions) GetPrometheus() *PrometheusOptions {
 	if x != nil {
 		return x.Prometheus
+	}
+	return nil
+}
+
+type TelemetryOptions struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Enabled              *ConfigurationVariable `protobuf:"bytes,1,opt,name=enabled,proto3" json:"enabled,omitempty"`
+	ExporterHttpEndpoint *ConfigurationVariable `protobuf:"bytes,2,opt,name=exporterHttpEndpoint,proto3" json:"exporterHttpEndpoint,omitempty"`
+	Sampler              *ConfigurationVariable `protobuf:"bytes,3,opt,name=sampler,proto3" json:"sampler,omitempty"`
+	AuthToken            *ConfigurationVariable `protobuf:"bytes,4,opt,name=authToken,proto3" json:"authToken,omitempty"`
+}
+
+func (x *TelemetryOptions) Reset() {
+	*x = TelemetryOptions{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_wundernode_config_proto_msgTypes[58]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *TelemetryOptions) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*TelemetryOptions) ProtoMessage() {}
+
+func (x *TelemetryOptions) ProtoReflect() protoreflect.Message {
+	mi := &file_wundernode_config_proto_msgTypes[58]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use TelemetryOptions.ProtoReflect.Descriptor instead.
+func (*TelemetryOptions) Descriptor() ([]byte, []int) {
+	return file_wundernode_config_proto_rawDescGZIP(), []int{58}
+}
+
+func (x *TelemetryOptions) GetEnabled() *ConfigurationVariable {
+	if x != nil {
+		return x.Enabled
+	}
+	return nil
+}
+
+func (x *TelemetryOptions) GetExporterHttpEndpoint() *ConfigurationVariable {
+	if x != nil {
+		return x.ExporterHttpEndpoint
+	}
+	return nil
+}
+
+func (x *TelemetryOptions) GetSampler() *ConfigurationVariable {
+	if x != nil {
+		return x.Sampler
+	}
+	return nil
+}
+
+func (x *TelemetryOptions) GetAuthToken() *ConfigurationVariable {
+	if x != nil {
+		return x.AuthToken
 	}
 	return nil
 }
@@ -4989,7 +5068,7 @@ type ServerLogging struct {
 func (x *ServerLogging) Reset() {
 	*x = ServerLogging{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_wundernode_config_proto_msgTypes[58]
+		mi := &file_wundernode_config_proto_msgTypes[59]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -5002,7 +5081,7 @@ func (x *ServerLogging) String() string {
 func (*ServerLogging) ProtoMessage() {}
 
 func (x *ServerLogging) ProtoReflect() protoreflect.Message {
-	mi := &file_wundernode_config_proto_msgTypes[58]
+	mi := &file_wundernode_config_proto_msgTypes[59]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -5015,7 +5094,7 @@ func (x *ServerLogging) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use ServerLogging.ProtoReflect.Descriptor instead.
 func (*ServerLogging) Descriptor() ([]byte, []int) {
-	return file_wundernode_config_proto_rawDescGZIP(), []int{58}
+	return file_wundernode_config_proto_rawDescGZIP(), []int{59}
 }
 
 func (x *ServerLogging) GetLevel() *ConfigurationVariable {
@@ -5038,7 +5117,7 @@ type ServerOptions struct {
 func (x *ServerOptions) Reset() {
 	*x = ServerOptions{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_wundernode_config_proto_msgTypes[59]
+		mi := &file_wundernode_config_proto_msgTypes[60]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -5051,7 +5130,7 @@ func (x *ServerOptions) String() string {
 func (*ServerOptions) ProtoMessage() {}
 
 func (x *ServerOptions) ProtoReflect() protoreflect.Message {
-	mi := &file_wundernode_config_proto_msgTypes[59]
+	mi := &file_wundernode_config_proto_msgTypes[60]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -5064,7 +5143,7 @@ func (x *ServerOptions) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use ServerOptions.ProtoReflect.Descriptor instead.
 func (*ServerOptions) Descriptor() ([]byte, []int) {
-	return file_wundernode_config_proto_rawDescGZIP(), []int{59}
+	return file_wundernode_config_proto_rawDescGZIP(), []int{60}
 }
 
 func (x *ServerOptions) GetServerUrl() *ConfigurationVariable {
@@ -5104,7 +5183,7 @@ type WebhookConfiguration struct {
 func (x *WebhookConfiguration) Reset() {
 	*x = WebhookConfiguration{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_wundernode_config_proto_msgTypes[60]
+		mi := &file_wundernode_config_proto_msgTypes[61]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -5117,7 +5196,7 @@ func (x *WebhookConfiguration) String() string {
 func (*WebhookConfiguration) ProtoMessage() {}
 
 func (x *WebhookConfiguration) ProtoReflect() protoreflect.Message {
-	mi := &file_wundernode_config_proto_msgTypes[60]
+	mi := &file_wundernode_config_proto_msgTypes[61]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -5130,7 +5209,7 @@ func (x *WebhookConfiguration) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use WebhookConfiguration.ProtoReflect.Descriptor instead.
 func (*WebhookConfiguration) Descriptor() ([]byte, []int) {
-	return file_wundernode_config_proto_rawDescGZIP(), []int{60}
+	return file_wundernode_config_proto_rawDescGZIP(), []int{61}
 }
 
 func (x *WebhookConfiguration) GetName() string {
@@ -5168,7 +5247,7 @@ type WebhookVerifier struct {
 func (x *WebhookVerifier) Reset() {
 	*x = WebhookVerifier{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_wundernode_config_proto_msgTypes[61]
+		mi := &file_wundernode_config_proto_msgTypes[62]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -5181,7 +5260,7 @@ func (x *WebhookVerifier) String() string {
 func (*WebhookVerifier) ProtoMessage() {}
 
 func (x *WebhookVerifier) ProtoReflect() protoreflect.Message {
-	mi := &file_wundernode_config_proto_msgTypes[61]
+	mi := &file_wundernode_config_proto_msgTypes[62]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -5194,7 +5273,7 @@ func (x *WebhookVerifier) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use WebhookVerifier.ProtoReflect.Descriptor instead.
 func (*WebhookVerifier) Descriptor() ([]byte, []int) {
-	return file_wundernode_config_proto_rawDescGZIP(), []int{61}
+	return file_wundernode_config_proto_rawDescGZIP(), []int{62}
 }
 
 func (x *WebhookVerifier) GetKind() WebhookVerifierKind {
@@ -5259,7 +5338,7 @@ type CorsConfiguration struct {
 func (x *CorsConfiguration) Reset() {
 	*x = CorsConfiguration{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_wundernode_config_proto_msgTypes[62]
+		mi := &file_wundernode_config_proto_msgTypes[63]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -5272,7 +5351,7 @@ func (x *CorsConfiguration) String() string {
 func (*CorsConfiguration) ProtoMessage() {}
 
 func (x *CorsConfiguration) ProtoReflect() protoreflect.Message {
-	mi := &file_wundernode_config_proto_msgTypes[62]
+	mi := &file_wundernode_config_proto_msgTypes[63]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -5285,7 +5364,7 @@ func (x *CorsConfiguration) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use CorsConfiguration.ProtoReflect.Descriptor instead.
 func (*CorsConfiguration) Descriptor() ([]byte, []int) {
-	return file_wundernode_config_proto_rawDescGZIP(), []int{62}
+	return file_wundernode_config_proto_rawDescGZIP(), []int{63}
 }
 
 func (x *CorsConfiguration) GetAllowedOrigins() []*ConfigurationVariable {
@@ -5345,7 +5424,7 @@ type ConfigurationVariable struct {
 func (x *ConfigurationVariable) Reset() {
 	*x = ConfigurationVariable{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_wundernode_config_proto_msgTypes[63]
+		mi := &file_wundernode_config_proto_msgTypes[64]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -5358,7 +5437,7 @@ func (x *ConfigurationVariable) String() string {
 func (*ConfigurationVariable) ProtoMessage() {}
 
 func (x *ConfigurationVariable) ProtoReflect() protoreflect.Message {
-	mi := &file_wundernode_config_proto_msgTypes[63]
+	mi := &file_wundernode_config_proto_msgTypes[64]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -5371,7 +5450,7 @@ func (x *ConfigurationVariable) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use ConfigurationVariable.ProtoReflect.Descriptor instead.
 func (*ConfigurationVariable) Descriptor() ([]byte, []int) {
-	return file_wundernode_config_proto_rawDescGZIP(), []int{63}
+	return file_wundernode_config_proto_rawDescGZIP(), []int{64}
 }
 
 func (x *ConfigurationVariable) GetKind() ConfigurationVariableKind {
@@ -5425,7 +5504,7 @@ type BuildInfo struct {
 func (x *BuildInfo) Reset() {
 	*x = BuildInfo{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_wundernode_config_proto_msgTypes[64]
+		mi := &file_wundernode_config_proto_msgTypes[65]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -5438,7 +5517,7 @@ func (x *BuildInfo) String() string {
 func (*BuildInfo) ProtoMessage() {}
 
 func (x *BuildInfo) ProtoReflect() protoreflect.Message {
-	mi := &file_wundernode_config_proto_msgTypes[64]
+	mi := &file_wundernode_config_proto_msgTypes[65]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -5451,7 +5530,7 @@ func (x *BuildInfo) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use BuildInfo.ProtoReflect.Descriptor instead.
 func (*BuildInfo) Descriptor() ([]byte, []int) {
-	return file_wundernode_config_proto_rawDescGZIP(), []int{64}
+	return file_wundernode_config_proto_rawDescGZIP(), []int{65}
 }
 
 func (x *BuildInfo) GetSuccess() bool {
@@ -5507,7 +5586,7 @@ type BuildInfoVersion struct {
 func (x *BuildInfoVersion) Reset() {
 	*x = BuildInfoVersion{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_wundernode_config_proto_msgTypes[65]
+		mi := &file_wundernode_config_proto_msgTypes[66]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -5520,7 +5599,7 @@ func (x *BuildInfoVersion) String() string {
 func (*BuildInfoVersion) ProtoMessage() {}
 
 func (x *BuildInfoVersion) ProtoReflect() protoreflect.Message {
-	mi := &file_wundernode_config_proto_msgTypes[65]
+	mi := &file_wundernode_config_proto_msgTypes[66]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -5533,7 +5612,7 @@ func (x *BuildInfoVersion) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use BuildInfoVersion.ProtoReflect.Descriptor instead.
 func (*BuildInfoVersion) Descriptor() ([]byte, []int) {
-	return file_wundernode_config_proto_rawDescGZIP(), []int{65}
+	return file_wundernode_config_proto_rawDescGZIP(), []int{66}
 }
 
 func (x *BuildInfoVersion) GetVersion() string {
@@ -5558,7 +5637,7 @@ type BuildInfoOS struct {
 func (x *BuildInfoOS) Reset() {
 	*x = BuildInfoOS{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_wundernode_config_proto_msgTypes[66]
+		mi := &file_wundernode_config_proto_msgTypes[67]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -5571,7 +5650,7 @@ func (x *BuildInfoOS) String() string {
 func (*BuildInfoOS) ProtoMessage() {}
 
 func (x *BuildInfoOS) ProtoReflect() protoreflect.Message {
-	mi := &file_wundernode_config_proto_msgTypes[66]
+	mi := &file_wundernode_config_proto_msgTypes[67]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -5584,7 +5663,7 @@ func (x *BuildInfoOS) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use BuildInfoOS.ProtoReflect.Descriptor instead.
 func (*BuildInfoOS) Descriptor() ([]byte, []int) {
-	return file_wundernode_config_proto_rawDescGZIP(), []int{66}
+	return file_wundernode_config_proto_rawDescGZIP(), []int{67}
 }
 
 func (x *BuildInfoOS) GetType() string {
@@ -5637,7 +5716,7 @@ type BuildInfoStats struct {
 func (x *BuildInfoStats) Reset() {
 	*x = BuildInfoStats{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_wundernode_config_proto_msgTypes[67]
+		mi := &file_wundernode_config_proto_msgTypes[68]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -5650,7 +5729,7 @@ func (x *BuildInfoStats) String() string {
 func (*BuildInfoStats) ProtoMessage() {}
 
 func (x *BuildInfoStats) ProtoReflect() protoreflect.Message {
-	mi := &file_wundernode_config_proto_msgTypes[67]
+	mi := &file_wundernode_config_proto_msgTypes[68]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -5663,7 +5742,7 @@ func (x *BuildInfoStats) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use BuildInfoStats.ProtoReflect.Descriptor instead.
 func (*BuildInfoStats) Descriptor() ([]byte, []int) {
-	return file_wundernode_config_proto_rawDescGZIP(), []int{67}
+	return file_wundernode_config_proto_rawDescGZIP(), []int{68}
 }
 
 func (x *BuildInfoStats) GetTotalApis() int32 {
@@ -6533,7 +6612,7 @@ var file_wundernode_config_proto_rawDesc = []byte{
 	0x65, 0x52, 0x07, 0x65, 0x6e, 0x61, 0x62, 0x6c, 0x65, 0x64, 0x12, 0x2f, 0x0a, 0x04, 0x70, 0x6f,
 	0x72, 0x74, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1b, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e,
 	0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x56, 0x61, 0x72,
-	0x69, 0x61, 0x62, 0x6c, 0x65, 0x52, 0x04, 0x70, 0x6f, 0x72, 0x74, 0x22, 0xbb, 0x04, 0x0a, 0x0b,
+	0x69, 0x61, 0x62, 0x6c, 0x65, 0x52, 0x04, 0x70, 0x6f, 0x72, 0x74, 0x22, 0xf9, 0x04, 0x0a, 0x0b,
 	0x4e, 0x6f, 0x64, 0x65, 0x4f, 0x70, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x12, 0x35, 0x0a, 0x07, 0x6e,
 	0x6f, 0x64, 0x65, 0x55, 0x72, 0x6c, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1b, 0x2e, 0x77,
 	0x67, 0x70, 0x62, 0x2e, 0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f,
@@ -6566,226 +6645,247 @@ var file_wundernode_config_proto_rawDesc = []byte{
 	0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74,
 	0x69, 0x6f, 0x6e, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x52, 0x13, 0x64, 0x65, 0x66,
 	0x61, 0x75, 0x6c, 0x74, 0x48, 0x74, 0x74, 0x70, 0x50, 0x72, 0x6f, 0x78, 0x79, 0x55, 0x72, 0x6c,
-	0x12, 0x37, 0x0a, 0x0a, 0x70, 0x72, 0x6f, 0x6d, 0x65, 0x74, 0x68, 0x65, 0x75, 0x73, 0x18, 0x09,
-	0x20, 0x01, 0x28, 0x0b, 0x32, 0x17, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x50, 0x72, 0x6f, 0x6d,
-	0x65, 0x74, 0x68, 0x65, 0x75, 0x73, 0x4f, 0x70, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x52, 0x0a, 0x70,
-	0x72, 0x6f, 0x6d, 0x65, 0x74, 0x68, 0x65, 0x75, 0x73, 0x22, 0x42, 0x0a, 0x0d, 0x53, 0x65, 0x72,
-	0x76, 0x65, 0x72, 0x4c, 0x6f, 0x67, 0x67, 0x69, 0x6e, 0x67, 0x12, 0x31, 0x0a, 0x05, 0x6c, 0x65,
-	0x76, 0x65, 0x6c, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1b, 0x2e, 0x77, 0x67, 0x70, 0x62,
-	0x2e, 0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x56, 0x61,
-	0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x52, 0x05, 0x6c, 0x65, 0x76, 0x65, 0x6c, 0x22, 0xa6, 0x01,
-	0x0a, 0x0d, 0x53, 0x65, 0x72, 0x76, 0x65, 0x72, 0x4f, 0x70, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x12,
-	0x39, 0x0a, 0x09, 0x73, 0x65, 0x72, 0x76, 0x65, 0x72, 0x55, 0x72, 0x6c, 0x18, 0x01, 0x20, 0x01,
-	0x28, 0x0b, 0x32, 0x1b, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67,
-	0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x52,
-	0x09, 0x73, 0x65, 0x72, 0x76, 0x65, 0x72, 0x55, 0x72, 0x6c, 0x12, 0x2d, 0x0a, 0x06, 0x6c, 0x69,
-	0x73, 0x74, 0x65, 0x6e, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x15, 0x2e, 0x77, 0x67, 0x70,
-	0x62, 0x2e, 0x4c, 0x69, 0x73, 0x74, 0x65, 0x6e, 0x65, 0x72, 0x4f, 0x70, 0x74, 0x69, 0x6f, 0x6e,
-	0x73, 0x52, 0x06, 0x6c, 0x69, 0x73, 0x74, 0x65, 0x6e, 0x12, 0x2b, 0x0a, 0x06, 0x6c, 0x6f, 0x67,
-	0x67, 0x65, 0x72, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x13, 0x2e, 0x77, 0x67, 0x70, 0x62,
-	0x2e, 0x53, 0x65, 0x72, 0x76, 0x65, 0x72, 0x4c, 0x6f, 0x67, 0x67, 0x69, 0x6e, 0x67, 0x52, 0x06,
-	0x6c, 0x6f, 0x67, 0x67, 0x65, 0x72, 0x22, 0x79, 0x0a, 0x14, 0x57, 0x65, 0x62, 0x68, 0x6f, 0x6f,
-	0x6b, 0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x12,
-	0x0a, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x6e, 0x61,
-	0x6d, 0x65, 0x12, 0x1a, 0x0a, 0x08, 0x66, 0x69, 0x6c, 0x65, 0x50, 0x61, 0x74, 0x68, 0x18, 0x02,
-	0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x66, 0x69, 0x6c, 0x65, 0x50, 0x61, 0x74, 0x68, 0x12, 0x31,
-	0x0a, 0x08, 0x76, 0x65, 0x72, 0x69, 0x66, 0x69, 0x65, 0x72, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b,
-	0x32, 0x15, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x57, 0x65, 0x62, 0x68, 0x6f, 0x6f, 0x6b, 0x56,
-	0x65, 0x72, 0x69, 0x66, 0x69, 0x65, 0x72, 0x52, 0x08, 0x76, 0x65, 0x72, 0x69, 0x66, 0x69, 0x65,
-	0x72, 0x22, 0xd5, 0x01, 0x0a, 0x0f, 0x57, 0x65, 0x62, 0x68, 0x6f, 0x6f, 0x6b, 0x56, 0x65, 0x72,
-	0x69, 0x66, 0x69, 0x65, 0x72, 0x12, 0x2d, 0x0a, 0x04, 0x6b, 0x69, 0x6e, 0x64, 0x18, 0x01, 0x20,
-	0x01, 0x28, 0x0e, 0x32, 0x19, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x57, 0x65, 0x62, 0x68, 0x6f,
-	0x6f, 0x6b, 0x56, 0x65, 0x72, 0x69, 0x66, 0x69, 0x65, 0x72, 0x4b, 0x69, 0x6e, 0x64, 0x52, 0x04,
-	0x6b, 0x69, 0x6e, 0x64, 0x12, 0x33, 0x0a, 0x06, 0x73, 0x65, 0x63, 0x72, 0x65, 0x74, 0x18, 0x02,
-	0x20, 0x01, 0x28, 0x0b, 0x32, 0x1b, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x43, 0x6f, 0x6e, 0x66,
+	0x12, 0x3c, 0x0a, 0x0d, 0x6f, 0x70, 0x65, 0x6e, 0x54, 0x65, 0x6c, 0x65, 0x6d, 0x65, 0x74, 0x72,
+	0x79, 0x18, 0x09, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x16, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x54,
+	0x65, 0x6c, 0x65, 0x6d, 0x65, 0x74, 0x72, 0x79, 0x4f, 0x70, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x52,
+	0x0d, 0x6f, 0x70, 0x65, 0x6e, 0x54, 0x65, 0x6c, 0x65, 0x6d, 0x65, 0x74, 0x72, 0x79, 0x12, 0x37,
+	0x0a, 0x0a, 0x70, 0x72, 0x6f, 0x6d, 0x65, 0x74, 0x68, 0x65, 0x75, 0x73, 0x18, 0x0a, 0x20, 0x01,
+	0x28, 0x0b, 0x32, 0x17, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x50, 0x72, 0x6f, 0x6d, 0x65, 0x74,
+	0x68, 0x65, 0x75, 0x73, 0x4f, 0x70, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x52, 0x0a, 0x70, 0x72, 0x6f,
+	0x6d, 0x65, 0x74, 0x68, 0x65, 0x75, 0x73, 0x22, 0x8c, 0x02, 0x0a, 0x10, 0x54, 0x65, 0x6c, 0x65,
+	0x6d, 0x65, 0x74, 0x72, 0x79, 0x4f, 0x70, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x12, 0x35, 0x0a, 0x07,
+	0x65, 0x6e, 0x61, 0x62, 0x6c, 0x65, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1b, 0x2e,
+	0x77, 0x67, 0x70, 0x62, 0x2e, 0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69,
+	0x6f, 0x6e, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x52, 0x07, 0x65, 0x6e, 0x61, 0x62,
+	0x6c, 0x65, 0x64, 0x12, 0x4f, 0x0a, 0x14, 0x65, 0x78, 0x70, 0x6f, 0x72, 0x74, 0x65, 0x72, 0x48,
+	0x74, 0x74, 0x70, 0x45, 0x6e, 0x64, 0x70, 0x6f, 0x69, 0x6e, 0x74, 0x18, 0x02, 0x20, 0x01, 0x28,
+	0x0b, 0x32, 0x1b, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x75,
+	0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x52, 0x14,
+	0x65, 0x78, 0x70, 0x6f, 0x72, 0x74, 0x65, 0x72, 0x48, 0x74, 0x74, 0x70, 0x45, 0x6e, 0x64, 0x70,
+	0x6f, 0x69, 0x6e, 0x74, 0x12, 0x35, 0x0a, 0x07, 0x73, 0x61, 0x6d, 0x70, 0x6c, 0x65, 0x72, 0x18,
+	0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1b, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x43, 0x6f, 0x6e,
+	0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62,
+	0x6c, 0x65, 0x52, 0x07, 0x73, 0x61, 0x6d, 0x70, 0x6c, 0x65, 0x72, 0x12, 0x39, 0x0a, 0x09, 0x61,
+	0x75, 0x74, 0x68, 0x54, 0x6f, 0x6b, 0x65, 0x6e, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1b,
+	0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74,
+	0x69, 0x6f, 0x6e, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x52, 0x09, 0x61, 0x75, 0x74,
+	0x68, 0x54, 0x6f, 0x6b, 0x65, 0x6e, 0x22, 0x42, 0x0a, 0x0d, 0x53, 0x65, 0x72, 0x76, 0x65, 0x72,
+	0x4c, 0x6f, 0x67, 0x67, 0x69, 0x6e, 0x67, 0x12, 0x31, 0x0a, 0x05, 0x6c, 0x65, 0x76, 0x65, 0x6c,
+	0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1b, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x43, 0x6f,
+	0x6e, 0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x56, 0x61, 0x72, 0x69, 0x61,
+	0x62, 0x6c, 0x65, 0x52, 0x05, 0x6c, 0x65, 0x76, 0x65, 0x6c, 0x22, 0xa6, 0x01, 0x0a, 0x0d, 0x53,
+	0x65, 0x72, 0x76, 0x65, 0x72, 0x4f, 0x70, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x12, 0x39, 0x0a, 0x09,
+	0x73, 0x65, 0x72, 0x76, 0x65, 0x72, 0x55, 0x72, 0x6c, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32,
+	0x1b, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x75, 0x72, 0x61,
+	0x74, 0x69, 0x6f, 0x6e, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x52, 0x09, 0x73, 0x65,
+	0x72, 0x76, 0x65, 0x72, 0x55, 0x72, 0x6c, 0x12, 0x2d, 0x0a, 0x06, 0x6c, 0x69, 0x73, 0x74, 0x65,
+	0x6e, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x15, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x4c,
+	0x69, 0x73, 0x74, 0x65, 0x6e, 0x65, 0x72, 0x4f, 0x70, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x52, 0x06,
+	0x6c, 0x69, 0x73, 0x74, 0x65, 0x6e, 0x12, 0x2b, 0x0a, 0x06, 0x6c, 0x6f, 0x67, 0x67, 0x65, 0x72,
+	0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x13, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x53, 0x65,
+	0x72, 0x76, 0x65, 0x72, 0x4c, 0x6f, 0x67, 0x67, 0x69, 0x6e, 0x67, 0x52, 0x06, 0x6c, 0x6f, 0x67,
+	0x67, 0x65, 0x72, 0x22, 0x79, 0x0a, 0x14, 0x57, 0x65, 0x62, 0x68, 0x6f, 0x6f, 0x6b, 0x43, 0x6f,
+	0x6e, 0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x12, 0x0a, 0x04, 0x6e,
+	0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x12,
+	0x1a, 0x0a, 0x08, 0x66, 0x69, 0x6c, 0x65, 0x50, 0x61, 0x74, 0x68, 0x18, 0x02, 0x20, 0x01, 0x28,
+	0x09, 0x52, 0x08, 0x66, 0x69, 0x6c, 0x65, 0x50, 0x61, 0x74, 0x68, 0x12, 0x31, 0x0a, 0x08, 0x76,
+	0x65, 0x72, 0x69, 0x66, 0x69, 0x65, 0x72, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x15, 0x2e,
+	0x77, 0x67, 0x70, 0x62, 0x2e, 0x57, 0x65, 0x62, 0x68, 0x6f, 0x6f, 0x6b, 0x56, 0x65, 0x72, 0x69,
+	0x66, 0x69, 0x65, 0x72, 0x52, 0x08, 0x76, 0x65, 0x72, 0x69, 0x66, 0x69, 0x65, 0x72, 0x22, 0xd5,
+	0x01, 0x0a, 0x0f, 0x57, 0x65, 0x62, 0x68, 0x6f, 0x6f, 0x6b, 0x56, 0x65, 0x72, 0x69, 0x66, 0x69,
+	0x65, 0x72, 0x12, 0x2d, 0x0a, 0x04, 0x6b, 0x69, 0x6e, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0e,
+	0x32, 0x19, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x57, 0x65, 0x62, 0x68, 0x6f, 0x6f, 0x6b, 0x56,
+	0x65, 0x72, 0x69, 0x66, 0x69, 0x65, 0x72, 0x4b, 0x69, 0x6e, 0x64, 0x52, 0x04, 0x6b, 0x69, 0x6e,
+	0x64, 0x12, 0x33, 0x0a, 0x06, 0x73, 0x65, 0x63, 0x72, 0x65, 0x74, 0x18, 0x02, 0x20, 0x01, 0x28,
+	0x0b, 0x32, 0x1b, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x75,
+	0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x52, 0x06,
+	0x73, 0x65, 0x63, 0x72, 0x65, 0x74, 0x12, 0x28, 0x0a, 0x0f, 0x73, 0x69, 0x67, 0x6e, 0x61, 0x74,
+	0x75, 0x72, 0x65, 0x48, 0x65, 0x61, 0x64, 0x65, 0x72, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52,
+	0x0f, 0x73, 0x69, 0x67, 0x6e, 0x61, 0x74, 0x75, 0x72, 0x65, 0x48, 0x65, 0x61, 0x64, 0x65, 0x72,
+	0x12, 0x34, 0x0a, 0x15, 0x73, 0x69, 0x67, 0x6e, 0x61, 0x74, 0x75, 0x72, 0x65, 0x48, 0x65, 0x61,
+	0x64, 0x65, 0x72, 0x50, 0x72, 0x65, 0x66, 0x69, 0x78, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52,
+	0x15, 0x73, 0x69, 0x67, 0x6e, 0x61, 0x74, 0x75, 0x72, 0x65, 0x48, 0x65, 0x61, 0x64, 0x65, 0x72,
+	0x50, 0x72, 0x65, 0x66, 0x69, 0x78, 0x22, 0x94, 0x02, 0x0a, 0x11, 0x43, 0x6f, 0x72, 0x73, 0x43,
+	0x6f, 0x6e, 0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x43, 0x0a, 0x0e,
+	0x61, 0x6c, 0x6c, 0x6f, 0x77, 0x65, 0x64, 0x4f, 0x72, 0x69, 0x67, 0x69, 0x6e, 0x73, 0x18, 0x01,
+	0x20, 0x03, 0x28, 0x0b, 0x32, 0x1b, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x43, 0x6f, 0x6e, 0x66,
 	0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c,
-	0x65, 0x52, 0x06, 0x73, 0x65, 0x63, 0x72, 0x65, 0x74, 0x12, 0x28, 0x0a, 0x0f, 0x73, 0x69, 0x67,
-	0x6e, 0x61, 0x74, 0x75, 0x72, 0x65, 0x48, 0x65, 0x61, 0x64, 0x65, 0x72, 0x18, 0x03, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x0f, 0x73, 0x69, 0x67, 0x6e, 0x61, 0x74, 0x75, 0x72, 0x65, 0x48, 0x65, 0x61,
-	0x64, 0x65, 0x72, 0x12, 0x34, 0x0a, 0x15, 0x73, 0x69, 0x67, 0x6e, 0x61, 0x74, 0x75, 0x72, 0x65,
-	0x48, 0x65, 0x61, 0x64, 0x65, 0x72, 0x50, 0x72, 0x65, 0x66, 0x69, 0x78, 0x18, 0x04, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x15, 0x73, 0x69, 0x67, 0x6e, 0x61, 0x74, 0x75, 0x72, 0x65, 0x48, 0x65, 0x61,
-	0x64, 0x65, 0x72, 0x50, 0x72, 0x65, 0x66, 0x69, 0x78, 0x22, 0x94, 0x02, 0x0a, 0x11, 0x43, 0x6f,
-	0x72, 0x73, 0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x12,
-	0x43, 0x0a, 0x0e, 0x61, 0x6c, 0x6c, 0x6f, 0x77, 0x65, 0x64, 0x4f, 0x72, 0x69, 0x67, 0x69, 0x6e,
-	0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x1b, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x43,
-	0x6f, 0x6e, 0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x56, 0x61, 0x72, 0x69,
-	0x61, 0x62, 0x6c, 0x65, 0x52, 0x0e, 0x61, 0x6c, 0x6c, 0x6f, 0x77, 0x65, 0x64, 0x4f, 0x72, 0x69,
-	0x67, 0x69, 0x6e, 0x73, 0x12, 0x26, 0x0a, 0x0e, 0x61, 0x6c, 0x6c, 0x6f, 0x77, 0x65, 0x64, 0x4d,
-	0x65, 0x74, 0x68, 0x6f, 0x64, 0x73, 0x18, 0x02, 0x20, 0x03, 0x28, 0x09, 0x52, 0x0e, 0x61, 0x6c,
-	0x6c, 0x6f, 0x77, 0x65, 0x64, 0x4d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x73, 0x12, 0x26, 0x0a, 0x0e,
-	0x61, 0x6c, 0x6c, 0x6f, 0x77, 0x65, 0x64, 0x48, 0x65, 0x61, 0x64, 0x65, 0x72, 0x73, 0x18, 0x03,
-	0x20, 0x03, 0x28, 0x09, 0x52, 0x0e, 0x61, 0x6c, 0x6c, 0x6f, 0x77, 0x65, 0x64, 0x48, 0x65, 0x61,
-	0x64, 0x65, 0x72, 0x73, 0x12, 0x26, 0x0a, 0x0e, 0x65, 0x78, 0x70, 0x6f, 0x73, 0x65, 0x64, 0x48,
-	0x65, 0x61, 0x64, 0x65, 0x72, 0x73, 0x18, 0x04, 0x20, 0x03, 0x28, 0x09, 0x52, 0x0e, 0x65, 0x78,
-	0x70, 0x6f, 0x73, 0x65, 0x64, 0x48, 0x65, 0x61, 0x64, 0x65, 0x72, 0x73, 0x12, 0x16, 0x0a, 0x06,
-	0x6d, 0x61, 0x78, 0x41, 0x67, 0x65, 0x18, 0x05, 0x20, 0x01, 0x28, 0x03, 0x52, 0x06, 0x6d, 0x61,
-	0x78, 0x41, 0x67, 0x65, 0x12, 0x2a, 0x0a, 0x10, 0x61, 0x6c, 0x6c, 0x6f, 0x77, 0x43, 0x72, 0x65,
-	0x64, 0x65, 0x6e, 0x74, 0x69, 0x61, 0x6c, 0x73, 0x18, 0x06, 0x20, 0x01, 0x28, 0x08, 0x52, 0x10,
-	0x61, 0x6c, 0x6c, 0x6f, 0x77, 0x43, 0x72, 0x65, 0x64, 0x65, 0x6e, 0x74, 0x69, 0x61, 0x6c, 0x73,
-	0x22, 0xc0, 0x02, 0x0a, 0x15, 0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69,
-	0x6f, 0x6e, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x12, 0x33, 0x0a, 0x04, 0x6b, 0x69,
-	0x6e, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0e, 0x32, 0x1f, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e,
-	0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x56, 0x61, 0x72,
-	0x69, 0x61, 0x62, 0x6c, 0x65, 0x4b, 0x69, 0x6e, 0x64, 0x52, 0x04, 0x6b, 0x69, 0x6e, 0x64, 0x12,
-	0x34, 0x0a, 0x15, 0x73, 0x74, 0x61, 0x74, 0x69, 0x63, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c,
-	0x65, 0x43, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x15,
+	0x65, 0x52, 0x0e, 0x61, 0x6c, 0x6c, 0x6f, 0x77, 0x65, 0x64, 0x4f, 0x72, 0x69, 0x67, 0x69, 0x6e,
+	0x73, 0x12, 0x26, 0x0a, 0x0e, 0x61, 0x6c, 0x6c, 0x6f, 0x77, 0x65, 0x64, 0x4d, 0x65, 0x74, 0x68,
+	0x6f, 0x64, 0x73, 0x18, 0x02, 0x20, 0x03, 0x28, 0x09, 0x52, 0x0e, 0x61, 0x6c, 0x6c, 0x6f, 0x77,
+	0x65, 0x64, 0x4d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x73, 0x12, 0x26, 0x0a, 0x0e, 0x61, 0x6c, 0x6c,
+	0x6f, 0x77, 0x65, 0x64, 0x48, 0x65, 0x61, 0x64, 0x65, 0x72, 0x73, 0x18, 0x03, 0x20, 0x03, 0x28,
+	0x09, 0x52, 0x0e, 0x61, 0x6c, 0x6c, 0x6f, 0x77, 0x65, 0x64, 0x48, 0x65, 0x61, 0x64, 0x65, 0x72,
+	0x73, 0x12, 0x26, 0x0a, 0x0e, 0x65, 0x78, 0x70, 0x6f, 0x73, 0x65, 0x64, 0x48, 0x65, 0x61, 0x64,
+	0x65, 0x72, 0x73, 0x18, 0x04, 0x20, 0x03, 0x28, 0x09, 0x52, 0x0e, 0x65, 0x78, 0x70, 0x6f, 0x73,
+	0x65, 0x64, 0x48, 0x65, 0x61, 0x64, 0x65, 0x72, 0x73, 0x12, 0x16, 0x0a, 0x06, 0x6d, 0x61, 0x78,
+	0x41, 0x67, 0x65, 0x18, 0x05, 0x20, 0x01, 0x28, 0x03, 0x52, 0x06, 0x6d, 0x61, 0x78, 0x41, 0x67,
+	0x65, 0x12, 0x2a, 0x0a, 0x10, 0x61, 0x6c, 0x6c, 0x6f, 0x77, 0x43, 0x72, 0x65, 0x64, 0x65, 0x6e,
+	0x74, 0x69, 0x61, 0x6c, 0x73, 0x18, 0x06, 0x20, 0x01, 0x28, 0x08, 0x52, 0x10, 0x61, 0x6c, 0x6c,
+	0x6f, 0x77, 0x43, 0x72, 0x65, 0x64, 0x65, 0x6e, 0x74, 0x69, 0x61, 0x6c, 0x73, 0x22, 0xc0, 0x02,
+	0x0a, 0x15, 0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x56,
+	0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x12, 0x33, 0x0a, 0x04, 0x6b, 0x69, 0x6e, 0x64, 0x18,
+	0x01, 0x20, 0x01, 0x28, 0x0e, 0x32, 0x1f, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x43, 0x6f, 0x6e,
+	0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62,
+	0x6c, 0x65, 0x4b, 0x69, 0x6e, 0x64, 0x52, 0x04, 0x6b, 0x69, 0x6e, 0x64, 0x12, 0x34, 0x0a, 0x15,
 	0x73, 0x74, 0x61, 0x74, 0x69, 0x63, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x43, 0x6f,
-	0x6e, 0x74, 0x65, 0x6e, 0x74, 0x12, 0x38, 0x0a, 0x17, 0x65, 0x6e, 0x76, 0x69, 0x72, 0x6f, 0x6e,
-	0x6d, 0x65, 0x6e, 0x74, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x4e, 0x61, 0x6d, 0x65,
-	0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x17, 0x65, 0x6e, 0x76, 0x69, 0x72, 0x6f, 0x6e, 0x6d,
-	0x65, 0x6e, 0x74, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x4e, 0x61, 0x6d, 0x65, 0x12,
-	0x48, 0x0a, 0x1f, 0x65, 0x6e, 0x76, 0x69, 0x72, 0x6f, 0x6e, 0x6d, 0x65, 0x6e, 0x74, 0x56, 0x61,
-	0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x44, 0x65, 0x66, 0x61, 0x75, 0x6c, 0x74, 0x56, 0x61, 0x6c,
-	0x75, 0x65, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x1f, 0x65, 0x6e, 0x76, 0x69, 0x72, 0x6f,
-	0x6e, 0x6d, 0x65, 0x6e, 0x74, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x44, 0x65, 0x66,
-	0x61, 0x75, 0x6c, 0x74, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x12, 0x38, 0x0a, 0x17, 0x70, 0x6c, 0x61,
-	0x63, 0x65, 0x68, 0x6f, 0x6c, 0x64, 0x65, 0x72, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65,
-	0x4e, 0x61, 0x6d, 0x65, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52, 0x17, 0x70, 0x6c, 0x61, 0x63,
-	0x65, 0x68, 0x6f, 0x6c, 0x64, 0x65, 0x72, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x4e,
-	0x61, 0x6d, 0x65, 0x22, 0x80, 0x02, 0x0a, 0x09, 0x42, 0x75, 0x69, 0x6c, 0x64, 0x49, 0x6e, 0x66,
-	0x6f, 0x12, 0x18, 0x0a, 0x07, 0x73, 0x75, 0x63, 0x63, 0x65, 0x73, 0x73, 0x18, 0x01, 0x20, 0x01,
-	0x28, 0x08, 0x52, 0x07, 0x73, 0x75, 0x63, 0x63, 0x65, 0x73, 0x73, 0x12, 0x28, 0x0a, 0x03, 0x73,
-	0x64, 0x6b, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x16, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e,
-	0x42, 0x75, 0x69, 0x6c, 0x64, 0x49, 0x6e, 0x66, 0x6f, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e,
-	0x52, 0x03, 0x73, 0x64, 0x6b, 0x12, 0x34, 0x0a, 0x09, 0x77, 0x75, 0x6e, 0x64, 0x65, 0x72, 0x63,
-	0x74, 0x6c, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x16, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e,
-	0x42, 0x75, 0x69, 0x6c, 0x64, 0x49, 0x6e, 0x66, 0x6f, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e,
-	0x52, 0x09, 0x77, 0x75, 0x6e, 0x64, 0x65, 0x72, 0x63, 0x74, 0x6c, 0x12, 0x2a, 0x0a, 0x04, 0x6e,
-	0x6f, 0x64, 0x65, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x16, 0x2e, 0x77, 0x67, 0x70, 0x62,
-	0x2e, 0x42, 0x75, 0x69, 0x6c, 0x64, 0x49, 0x6e, 0x66, 0x6f, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f,
-	0x6e, 0x52, 0x04, 0x6e, 0x6f, 0x64, 0x65, 0x12, 0x21, 0x0a, 0x02, 0x6f, 0x73, 0x18, 0x05, 0x20,
-	0x01, 0x28, 0x0b, 0x32, 0x11, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x42, 0x75, 0x69, 0x6c, 0x64,
-	0x49, 0x6e, 0x66, 0x6f, 0x4f, 0x53, 0x52, 0x02, 0x6f, 0x73, 0x12, 0x2a, 0x0a, 0x05, 0x73, 0x74,
-	0x61, 0x74, 0x73, 0x18, 0x06, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x14, 0x2e, 0x77, 0x67, 0x70, 0x62,
-	0x2e, 0x42, 0x75, 0x69, 0x6c, 0x64, 0x49, 0x6e, 0x66, 0x6f, 0x53, 0x74, 0x61, 0x74, 0x73, 0x52,
-	0x05, 0x73, 0x74, 0x61, 0x74, 0x73, 0x22, 0x2c, 0x0a, 0x10, 0x42, 0x75, 0x69, 0x6c, 0x64, 0x49,
-	0x6e, 0x66, 0x6f, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x12, 0x18, 0x0a, 0x07, 0x76, 0x65,
-	0x72, 0x73, 0x69, 0x6f, 0x6e, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x76, 0x65, 0x72,
-	0x73, 0x69, 0x6f, 0x6e, 0x22, 0x85, 0x01, 0x0a, 0x0b, 0x42, 0x75, 0x69, 0x6c, 0x64, 0x49, 0x6e,
-	0x66, 0x6f, 0x4f, 0x53, 0x12, 0x12, 0x0a, 0x04, 0x74, 0x79, 0x70, 0x65, 0x18, 0x01, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x04, 0x74, 0x79, 0x70, 0x65, 0x12, 0x1a, 0x0a, 0x08, 0x70, 0x6c, 0x61, 0x74,
-	0x66, 0x6f, 0x72, 0x6d, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x70, 0x6c, 0x61, 0x74,
-	0x66, 0x6f, 0x72, 0x6d, 0x12, 0x12, 0x0a, 0x04, 0x61, 0x72, 0x63, 0x68, 0x18, 0x03, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x04, 0x61, 0x72, 0x63, 0x68, 0x12, 0x18, 0x0a, 0x07, 0x76, 0x65, 0x72, 0x73,
-	0x69, 0x6f, 0x6e, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x76, 0x65, 0x72, 0x73, 0x69,
-	0x6f, 0x6e, 0x12, 0x18, 0x0a, 0x07, 0x72, 0x65, 0x6c, 0x65, 0x61, 0x73, 0x65, 0x18, 0x05, 0x20,
-	0x01, 0x28, 0x09, 0x52, 0x07, 0x72, 0x65, 0x6c, 0x65, 0x61, 0x73, 0x65, 0x22, 0xea, 0x01, 0x0a,
-	0x0e, 0x42, 0x75, 0x69, 0x6c, 0x64, 0x49, 0x6e, 0x66, 0x6f, 0x53, 0x74, 0x61, 0x74, 0x73, 0x12,
-	0x1c, 0x0a, 0x09, 0x74, 0x6f, 0x74, 0x61, 0x6c, 0x41, 0x70, 0x69, 0x73, 0x18, 0x01, 0x20, 0x01,
-	0x28, 0x05, 0x52, 0x09, 0x74, 0x6f, 0x74, 0x61, 0x6c, 0x41, 0x70, 0x69, 0x73, 0x12, 0x28, 0x0a,
-	0x0f, 0x74, 0x6f, 0x74, 0x61, 0x6c, 0x4f, 0x70, 0x65, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x73,
-	0x18, 0x02, 0x20, 0x01, 0x28, 0x05, 0x52, 0x0f, 0x74, 0x6f, 0x74, 0x61, 0x6c, 0x4f, 0x70, 0x65,
-	0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x12, 0x24, 0x0a, 0x0d, 0x74, 0x6f, 0x74, 0x61, 0x6c,
-	0x57, 0x65, 0x62, 0x68, 0x6f, 0x6f, 0x6b, 0x73, 0x18, 0x03, 0x20, 0x01, 0x28, 0x05, 0x52, 0x0d,
-	0x74, 0x6f, 0x74, 0x61, 0x6c, 0x57, 0x65, 0x62, 0x68, 0x6f, 0x6f, 0x6b, 0x73, 0x12, 0x3c, 0x0a,
-	0x19, 0x68, 0x61, 0x73, 0x41, 0x75, 0x74, 0x68, 0x65, 0x6e, 0x74, 0x69, 0x63, 0x61, 0x74, 0x69,
-	0x6f, 0x6e, 0x50, 0x72, 0x6f, 0x76, 0x69, 0x64, 0x65, 0x72, 0x18, 0x04, 0x20, 0x01, 0x28, 0x08,
-	0x52, 0x19, 0x68, 0x61, 0x73, 0x41, 0x75, 0x74, 0x68, 0x65, 0x6e, 0x74, 0x69, 0x63, 0x61, 0x74,
-	0x69, 0x6f, 0x6e, 0x50, 0x72, 0x6f, 0x76, 0x69, 0x64, 0x65, 0x72, 0x12, 0x2c, 0x0a, 0x11, 0x68,
-	0x61, 0x73, 0x55, 0x70, 0x6c, 0x6f, 0x61, 0x64, 0x50, 0x72, 0x6f, 0x76, 0x69, 0x64, 0x65, 0x72,
-	0x18, 0x05, 0x20, 0x01, 0x28, 0x08, 0x52, 0x11, 0x68, 0x61, 0x73, 0x55, 0x70, 0x6c, 0x6f, 0x61,
-	0x64, 0x50, 0x72, 0x6f, 0x76, 0x69, 0x64, 0x65, 0x72, 0x2a, 0x4d, 0x0a, 0x08, 0x4c, 0x6f, 0x67,
-	0x4c, 0x65, 0x76, 0x65, 0x6c, 0x12, 0x09, 0x0a, 0x05, 0x44, 0x45, 0x42, 0x55, 0x47, 0x10, 0x00,
-	0x12, 0x08, 0x0a, 0x04, 0x49, 0x4e, 0x46, 0x4f, 0x10, 0x01, 0x12, 0x09, 0x0a, 0x05, 0x45, 0x52,
-	0x52, 0x4f, 0x52, 0x10, 0x02, 0x12, 0x0b, 0x0a, 0x07, 0x57, 0x41, 0x52, 0x4e, 0x49, 0x4e, 0x47,
-	0x10, 0x03, 0x12, 0x09, 0x0a, 0x05, 0x50, 0x41, 0x4e, 0x49, 0x43, 0x10, 0x04, 0x12, 0x09, 0x0a,
-	0x05, 0x46, 0x41, 0x54, 0x41, 0x4c, 0x10, 0x05, 0x2a, 0x57, 0x0a, 0x10, 0x41, 0x75, 0x74, 0x68,
-	0x50, 0x72, 0x6f, 0x76, 0x69, 0x64, 0x65, 0x72, 0x4b, 0x69, 0x6e, 0x64, 0x12, 0x16, 0x0a, 0x12,
-	0x41, 0x75, 0x74, 0x68, 0x50, 0x72, 0x6f, 0x76, 0x69, 0x64, 0x65, 0x72, 0x47, 0x69, 0x74, 0x68,
-	0x75, 0x62, 0x10, 0x00, 0x12, 0x14, 0x0a, 0x10, 0x41, 0x75, 0x74, 0x68, 0x50, 0x72, 0x6f, 0x76,
-	0x69, 0x64, 0x65, 0x72, 0x4f, 0x49, 0x44, 0x43, 0x10, 0x01, 0x12, 0x15, 0x0a, 0x11, 0x41, 0x75,
-	0x74, 0x68, 0x50, 0x72, 0x6f, 0x76, 0x69, 0x64, 0x65, 0x72, 0x41, 0x75, 0x74, 0x68, 0x30, 0x10,
-	0x02, 0x2a, 0x41, 0x0a, 0x18, 0x4f, 0x70, 0x65, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x45, 0x78,
-	0x65, 0x63, 0x75, 0x74, 0x69, 0x6f, 0x6e, 0x45, 0x6e, 0x67, 0x69, 0x6e, 0x65, 0x12, 0x12, 0x0a,
-	0x0e, 0x45, 0x4e, 0x47, 0x49, 0x4e, 0x45, 0x5f, 0x47, 0x52, 0x41, 0x50, 0x48, 0x51, 0x4c, 0x10,
-	0x00, 0x12, 0x11, 0x0a, 0x0d, 0x45, 0x4e, 0x47, 0x49, 0x4e, 0x45, 0x5f, 0x4e, 0x4f, 0x44, 0x45,
-	0x4a, 0x53, 0x10, 0x01, 0x2a, 0x44, 0x0a, 0x1d, 0x50, 0x6f, 0x73, 0x74, 0x52, 0x65, 0x73, 0x6f,
-	0x6c, 0x76, 0x65, 0x54, 0x72, 0x61, 0x6e, 0x73, 0x66, 0x6f, 0x72, 0x6d, 0x61, 0x74, 0x69, 0x6f,
-	0x6e, 0x4b, 0x69, 0x6e, 0x64, 0x12, 0x23, 0x0a, 0x1f, 0x47, 0x45, 0x54, 0x5f, 0x50, 0x4f, 0x53,
-	0x54, 0x5f, 0x52, 0x45, 0x53, 0x4f, 0x4c, 0x56, 0x45, 0x5f, 0x54, 0x52, 0x41, 0x4e, 0x53, 0x46,
-	0x4f, 0x52, 0x4d, 0x41, 0x54, 0x49, 0x4f, 0x4e, 0x10, 0x00, 0x2a, 0x47, 0x0a, 0x12, 0x49, 0x6e,
-	0x6a, 0x65, 0x63, 0x74, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x4b, 0x69, 0x6e, 0x64,
-	0x12, 0x08, 0x0a, 0x04, 0x55, 0x55, 0x49, 0x44, 0x10, 0x00, 0x12, 0x0d, 0x0a, 0x09, 0x44, 0x41,
-	0x54, 0x45, 0x5f, 0x54, 0x49, 0x4d, 0x45, 0x10, 0x01, 0x12, 0x18, 0x0a, 0x14, 0x45, 0x4e, 0x56,
-	0x49, 0x52, 0x4f, 0x4e, 0x4d, 0x45, 0x4e, 0x54, 0x5f, 0x56, 0x41, 0x52, 0x49, 0x41, 0x42, 0x4c,
-	0x45, 0x10, 0x02, 0x2a, 0xbc, 0x02, 0x0a, 0x09, 0x43, 0x6c, 0x61, 0x69, 0x6d, 0x54, 0x79, 0x70,
-	0x65, 0x12, 0x0a, 0x0a, 0x06, 0x49, 0x53, 0x53, 0x55, 0x45, 0x52, 0x10, 0x00, 0x12, 0x0c, 0x0a,
-	0x08, 0x50, 0x52, 0x4f, 0x56, 0x49, 0x44, 0x45, 0x52, 0x10, 0x00, 0x12, 0x0b, 0x0a, 0x07, 0x53,
-	0x55, 0x42, 0x4a, 0x45, 0x43, 0x54, 0x10, 0x01, 0x12, 0x0a, 0x0a, 0x06, 0x55, 0x53, 0x45, 0x52,
-	0x49, 0x44, 0x10, 0x01, 0x12, 0x08, 0x0a, 0x04, 0x4e, 0x41, 0x4d, 0x45, 0x10, 0x02, 0x12, 0x0e,
-	0x0a, 0x0a, 0x47, 0x49, 0x56, 0x45, 0x4e, 0x5f, 0x4e, 0x41, 0x4d, 0x45, 0x10, 0x03, 0x12, 0x0f,
-	0x0a, 0x0b, 0x46, 0x41, 0x4d, 0x49, 0x4c, 0x59, 0x5f, 0x4e, 0x41, 0x4d, 0x45, 0x10, 0x04, 0x12,
-	0x0f, 0x0a, 0x0b, 0x4d, 0x49, 0x44, 0x44, 0x4c, 0x45, 0x5f, 0x4e, 0x41, 0x4d, 0x45, 0x10, 0x05,
-	0x12, 0x0c, 0x0a, 0x08, 0x4e, 0x49, 0x43, 0x4b, 0x4e, 0x41, 0x4d, 0x45, 0x10, 0x06, 0x12, 0x16,
-	0x0a, 0x12, 0x50, 0x52, 0x45, 0x46, 0x45, 0x52, 0x52, 0x45, 0x44, 0x5f, 0x55, 0x53, 0x45, 0x52,
-	0x4e, 0x41, 0x4d, 0x45, 0x10, 0x07, 0x12, 0x0b, 0x0a, 0x07, 0x50, 0x52, 0x4f, 0x46, 0x49, 0x4c,
-	0x45, 0x10, 0x08, 0x12, 0x0b, 0x0a, 0x07, 0x50, 0x49, 0x43, 0x54, 0x55, 0x52, 0x45, 0x10, 0x09,
-	0x12, 0x0b, 0x0a, 0x07, 0x57, 0x45, 0x42, 0x53, 0x49, 0x54, 0x45, 0x10, 0x0a, 0x12, 0x09, 0x0a,
-	0x05, 0x45, 0x4d, 0x41, 0x49, 0x4c, 0x10, 0x0b, 0x12, 0x12, 0x0a, 0x0e, 0x45, 0x4d, 0x41, 0x49,
-	0x4c, 0x5f, 0x56, 0x45, 0x52, 0x49, 0x46, 0x49, 0x45, 0x44, 0x10, 0x0c, 0x12, 0x0a, 0x0a, 0x06,
-	0x47, 0x45, 0x4e, 0x44, 0x45, 0x52, 0x10, 0x0d, 0x12, 0x0e, 0x0a, 0x0a, 0x42, 0x49, 0x52, 0x54,
-	0x48, 0x5f, 0x44, 0x41, 0x54, 0x45, 0x10, 0x0e, 0x12, 0x0d, 0x0a, 0x09, 0x5a, 0x4f, 0x4e, 0x45,
-	0x5f, 0x49, 0x4e, 0x46, 0x4f, 0x10, 0x0f, 0x12, 0x0a, 0x0a, 0x06, 0x4c, 0x4f, 0x43, 0x41, 0x4c,
-	0x45, 0x10, 0x10, 0x12, 0x0c, 0x0a, 0x08, 0x4c, 0x4f, 0x43, 0x41, 0x54, 0x49, 0x4f, 0x4e, 0x10,
-	0x11, 0x12, 0x0b, 0x0a, 0x06, 0x43, 0x55, 0x53, 0x54, 0x4f, 0x4d, 0x10, 0xe7, 0x07, 0x1a, 0x02,
-	0x10, 0x01, 0x2a, 0x41, 0x0a, 0x09, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x54, 0x79, 0x70, 0x65, 0x12,
-	0x0a, 0x0a, 0x06, 0x53, 0x54, 0x52, 0x49, 0x4e, 0x47, 0x10, 0x00, 0x12, 0x07, 0x0a, 0x03, 0x49,
-	0x4e, 0x54, 0x10, 0x01, 0x12, 0x09, 0x0a, 0x05, 0x46, 0x4c, 0x4f, 0x41, 0x54, 0x10, 0x02, 0x12,
-	0x0b, 0x0a, 0x07, 0x42, 0x4f, 0x4f, 0x4c, 0x45, 0x41, 0x4e, 0x10, 0x03, 0x12, 0x07, 0x0a, 0x03,
-	0x41, 0x4e, 0x59, 0x10, 0x04, 0x2a, 0x3a, 0x0a, 0x0d, 0x4f, 0x70, 0x65, 0x72, 0x61, 0x74, 0x69,
-	0x6f, 0x6e, 0x54, 0x79, 0x70, 0x65, 0x12, 0x09, 0x0a, 0x05, 0x51, 0x55, 0x45, 0x52, 0x59, 0x10,
-	0x00, 0x12, 0x0c, 0x0a, 0x08, 0x4d, 0x55, 0x54, 0x41, 0x54, 0x49, 0x4f, 0x4e, 0x10, 0x01, 0x12,
-	0x10, 0x0a, 0x0c, 0x53, 0x55, 0x42, 0x53, 0x43, 0x52, 0x49, 0x50, 0x54, 0x49, 0x4f, 0x4e, 0x10,
-	0x02, 0x2a, 0x82, 0x01, 0x0a, 0x0e, 0x44, 0x61, 0x74, 0x61, 0x53, 0x6f, 0x75, 0x72, 0x63, 0x65,
-	0x4b, 0x69, 0x6e, 0x64, 0x12, 0x0a, 0x0a, 0x06, 0x53, 0x54, 0x41, 0x54, 0x49, 0x43, 0x10, 0x00,
-	0x12, 0x08, 0x0a, 0x04, 0x52, 0x45, 0x53, 0x54, 0x10, 0x01, 0x12, 0x0b, 0x0a, 0x07, 0x47, 0x52,
-	0x41, 0x50, 0x48, 0x51, 0x4c, 0x10, 0x02, 0x12, 0x0e, 0x0a, 0x0a, 0x50, 0x4f, 0x53, 0x54, 0x47,
-	0x52, 0x45, 0x53, 0x51, 0x4c, 0x10, 0x03, 0x12, 0x09, 0x0a, 0x05, 0x4d, 0x59, 0x53, 0x51, 0x4c,
-	0x10, 0x04, 0x12, 0x0d, 0x0a, 0x09, 0x53, 0x51, 0x4c, 0x53, 0x45, 0x52, 0x56, 0x45, 0x52, 0x10,
-	0x05, 0x12, 0x0b, 0x0a, 0x07, 0x4d, 0x4f, 0x4e, 0x47, 0x4f, 0x44, 0x42, 0x10, 0x06, 0x12, 0x0a,
-	0x0a, 0x06, 0x53, 0x51, 0x4c, 0x49, 0x54, 0x45, 0x10, 0x07, 0x12, 0x0a, 0x0a, 0x06, 0x50, 0x52,
-	0x49, 0x53, 0x4d, 0x41, 0x10, 0x08, 0x2a, 0x71, 0x0a, 0x1a, 0x55, 0x70, 0x73, 0x74, 0x72, 0x65,
-	0x61, 0x6d, 0x41, 0x75, 0x74, 0x68, 0x65, 0x6e, 0x74, 0x69, 0x63, 0x61, 0x74, 0x69, 0x6f, 0x6e,
-	0x4b, 0x69, 0x6e, 0x64, 0x12, 0x1d, 0x0a, 0x19, 0x55, 0x70, 0x73, 0x74, 0x72, 0x65, 0x61, 0x6d,
-	0x41, 0x75, 0x74, 0x68, 0x65, 0x6e, 0x74, 0x69, 0x63, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x4a, 0x57,
-	0x54, 0x10, 0x00, 0x12, 0x34, 0x0a, 0x30, 0x55, 0x70, 0x73, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x41,
-	0x75, 0x74, 0x68, 0x65, 0x6e, 0x74, 0x69, 0x63, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x4a, 0x57, 0x54,
-	0x57, 0x69, 0x74, 0x68, 0x41, 0x63, 0x63, 0x65, 0x73, 0x73, 0x54, 0x6f, 0x6b, 0x65, 0x6e, 0x45,
-	0x78, 0x63, 0x68, 0x61, 0x6e, 0x67, 0x65, 0x10, 0x01, 0x2a, 0x27, 0x0a, 0x0d, 0x53, 0x69, 0x67,
-	0x6e, 0x69, 0x6e, 0x67, 0x4d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x12, 0x16, 0x0a, 0x12, 0x53, 0x69,
-	0x67, 0x6e, 0x69, 0x6e, 0x67, 0x4d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x48, 0x53, 0x32, 0x35, 0x36,
-	0x10, 0x00, 0x2a, 0x41, 0x0a, 0x0a, 0x48, 0x54, 0x54, 0x50, 0x4d, 0x65, 0x74, 0x68, 0x6f, 0x64,
-	0x12, 0x07, 0x0a, 0x03, 0x47, 0x45, 0x54, 0x10, 0x00, 0x12, 0x08, 0x0a, 0x04, 0x50, 0x4f, 0x53,
-	0x54, 0x10, 0x01, 0x12, 0x07, 0x0a, 0x03, 0x50, 0x55, 0x54, 0x10, 0x02, 0x12, 0x0a, 0x0a, 0x06,
-	0x44, 0x45, 0x4c, 0x45, 0x54, 0x45, 0x10, 0x03, 0x12, 0x0b, 0x0a, 0x07, 0x4f, 0x50, 0x54, 0x49,
-	0x4f, 0x4e, 0x53, 0x10, 0x04, 0x2a, 0x36, 0x0a, 0x0e, 0x41, 0x72, 0x67, 0x75, 0x6d, 0x65, 0x6e,
-	0x74, 0x53, 0x6f, 0x75, 0x72, 0x63, 0x65, 0x12, 0x10, 0x0a, 0x0c, 0x4f, 0x42, 0x4a, 0x45, 0x43,
-	0x54, 0x5f, 0x46, 0x49, 0x45, 0x4c, 0x44, 0x10, 0x00, 0x12, 0x12, 0x0a, 0x0e, 0x46, 0x49, 0x45,
-	0x4c, 0x44, 0x5f, 0x41, 0x52, 0x47, 0x55, 0x4d, 0x45, 0x4e, 0x54, 0x10, 0x01, 0x2a, 0x82, 0x01,
-	0x0a, 0x1b, 0x41, 0x72, 0x67, 0x75, 0x6d, 0x65, 0x6e, 0x74, 0x52, 0x65, 0x6e, 0x64, 0x65, 0x72,
-	0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x1b, 0x0a,
-	0x17, 0x52, 0x45, 0x4e, 0x44, 0x45, 0x52, 0x5f, 0x41, 0x52, 0x47, 0x55, 0x4d, 0x45, 0x4e, 0x54,
-	0x5f, 0x44, 0x45, 0x46, 0x41, 0x55, 0x4c, 0x54, 0x10, 0x00, 0x12, 0x24, 0x0a, 0x20, 0x52, 0x45,
-	0x4e, 0x44, 0x45, 0x52, 0x5f, 0x41, 0x52, 0x47, 0x55, 0x4d, 0x45, 0x4e, 0x54, 0x5f, 0x41, 0x53,
-	0x5f, 0x47, 0x52, 0x41, 0x50, 0x48, 0x51, 0x4c, 0x5f, 0x56, 0x41, 0x4c, 0x55, 0x45, 0x10, 0x01,
-	0x12, 0x20, 0x0a, 0x1c, 0x52, 0x45, 0x4e, 0x44, 0x45, 0x52, 0x5f, 0x41, 0x52, 0x47, 0x55, 0x4d,
-	0x45, 0x4e, 0x54, 0x5f, 0x41, 0x53, 0x5f, 0x41, 0x52, 0x52, 0x41, 0x59, 0x5f, 0x43, 0x53, 0x56,
-	0x10, 0x02, 0x2a, 0x26, 0x0a, 0x13, 0x57, 0x65, 0x62, 0x68, 0x6f, 0x6f, 0x6b, 0x56, 0x65, 0x72,
-	0x69, 0x66, 0x69, 0x65, 0x72, 0x4b, 0x69, 0x6e, 0x64, 0x12, 0x0f, 0x0a, 0x0b, 0x48, 0x4d, 0x41,
-	0x43, 0x5f, 0x53, 0x48, 0x41, 0x32, 0x35, 0x36, 0x10, 0x00, 0x2a, 0x86, 0x01, 0x0a, 0x19, 0x43,
-	0x6f, 0x6e, 0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x56, 0x61, 0x72, 0x69,
-	0x61, 0x62, 0x6c, 0x65, 0x4b, 0x69, 0x6e, 0x64, 0x12, 0x21, 0x0a, 0x1d, 0x53, 0x54, 0x41, 0x54,
-	0x49, 0x43, 0x5f, 0x43, 0x4f, 0x4e, 0x46, 0x49, 0x47, 0x55, 0x52, 0x41, 0x54, 0x49, 0x4f, 0x4e,
-	0x5f, 0x56, 0x41, 0x52, 0x49, 0x41, 0x42, 0x4c, 0x45, 0x10, 0x00, 0x12, 0x1e, 0x0a, 0x1a, 0x45,
-	0x4e, 0x56, 0x5f, 0x43, 0x4f, 0x4e, 0x46, 0x49, 0x47, 0x55, 0x52, 0x41, 0x54, 0x49, 0x4f, 0x4e,
-	0x5f, 0x56, 0x41, 0x52, 0x49, 0x41, 0x42, 0x4c, 0x45, 0x10, 0x01, 0x12, 0x26, 0x0a, 0x22, 0x50,
-	0x4c, 0x41, 0x43, 0x45, 0x48, 0x4f, 0x4c, 0x44, 0x45, 0x52, 0x5f, 0x43, 0x4f, 0x4e, 0x46, 0x49,
-	0x47, 0x55, 0x52, 0x41, 0x54, 0x49, 0x4f, 0x4e, 0x5f, 0x56, 0x41, 0x52, 0x49, 0x41, 0x42, 0x4c,
-	0x45, 0x10, 0x02, 0x42, 0x2d, 0x5a, 0x2b, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f,
-	0x6d, 0x2f, 0x77, 0x75, 0x6e, 0x64, 0x65, 0x72, 0x67, 0x72, 0x61, 0x70, 0x68, 0x2f, 0x77, 0x75,
-	0x6e, 0x64, 0x65, 0x72, 0x67, 0x72, 0x61, 0x70, 0x68, 0x2f, 0x70, 0x6b, 0x67, 0x2f, 0x77, 0x67,
-	0x70, 0x62, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x6e, 0x74, 0x65, 0x6e, 0x74, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x15, 0x73, 0x74, 0x61,
+	0x74, 0x69, 0x63, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x43, 0x6f, 0x6e, 0x74, 0x65,
+	0x6e, 0x74, 0x12, 0x38, 0x0a, 0x17, 0x65, 0x6e, 0x76, 0x69, 0x72, 0x6f, 0x6e, 0x6d, 0x65, 0x6e,
+	0x74, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x4e, 0x61, 0x6d, 0x65, 0x18, 0x03, 0x20,
+	0x01, 0x28, 0x09, 0x52, 0x17, 0x65, 0x6e, 0x76, 0x69, 0x72, 0x6f, 0x6e, 0x6d, 0x65, 0x6e, 0x74,
+	0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x48, 0x0a, 0x1f,
+	0x65, 0x6e, 0x76, 0x69, 0x72, 0x6f, 0x6e, 0x6d, 0x65, 0x6e, 0x74, 0x56, 0x61, 0x72, 0x69, 0x61,
+	0x62, 0x6c, 0x65, 0x44, 0x65, 0x66, 0x61, 0x75, 0x6c, 0x74, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x18,
+	0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x1f, 0x65, 0x6e, 0x76, 0x69, 0x72, 0x6f, 0x6e, 0x6d, 0x65,
+	0x6e, 0x74, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x44, 0x65, 0x66, 0x61, 0x75, 0x6c,
+	0x74, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x12, 0x38, 0x0a, 0x17, 0x70, 0x6c, 0x61, 0x63, 0x65, 0x68,
+	0x6f, 0x6c, 0x64, 0x65, 0x72, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x4e, 0x61, 0x6d,
+	0x65, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52, 0x17, 0x70, 0x6c, 0x61, 0x63, 0x65, 0x68, 0x6f,
+	0x6c, 0x64, 0x65, 0x72, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x4e, 0x61, 0x6d, 0x65,
+	0x22, 0x80, 0x02, 0x0a, 0x09, 0x42, 0x75, 0x69, 0x6c, 0x64, 0x49, 0x6e, 0x66, 0x6f, 0x12, 0x18,
+	0x0a, 0x07, 0x73, 0x75, 0x63, 0x63, 0x65, 0x73, 0x73, 0x18, 0x01, 0x20, 0x01, 0x28, 0x08, 0x52,
+	0x07, 0x73, 0x75, 0x63, 0x63, 0x65, 0x73, 0x73, 0x12, 0x28, 0x0a, 0x03, 0x73, 0x64, 0x6b, 0x18,
+	0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x16, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x42, 0x75, 0x69,
+	0x6c, 0x64, 0x49, 0x6e, 0x66, 0x6f, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x52, 0x03, 0x73,
+	0x64, 0x6b, 0x12, 0x34, 0x0a, 0x09, 0x77, 0x75, 0x6e, 0x64, 0x65, 0x72, 0x63, 0x74, 0x6c, 0x18,
+	0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x16, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x42, 0x75, 0x69,
+	0x6c, 0x64, 0x49, 0x6e, 0x66, 0x6f, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x52, 0x09, 0x77,
+	0x75, 0x6e, 0x64, 0x65, 0x72, 0x63, 0x74, 0x6c, 0x12, 0x2a, 0x0a, 0x04, 0x6e, 0x6f, 0x64, 0x65,
+	0x18, 0x04, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x16, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x42, 0x75,
+	0x69, 0x6c, 0x64, 0x49, 0x6e, 0x66, 0x6f, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x52, 0x04,
+	0x6e, 0x6f, 0x64, 0x65, 0x12, 0x21, 0x0a, 0x02, 0x6f, 0x73, 0x18, 0x05, 0x20, 0x01, 0x28, 0x0b,
+	0x32, 0x11, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x42, 0x75, 0x69, 0x6c, 0x64, 0x49, 0x6e, 0x66,
+	0x6f, 0x4f, 0x53, 0x52, 0x02, 0x6f, 0x73, 0x12, 0x2a, 0x0a, 0x05, 0x73, 0x74, 0x61, 0x74, 0x73,
+	0x18, 0x06, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x14, 0x2e, 0x77, 0x67, 0x70, 0x62, 0x2e, 0x42, 0x75,
+	0x69, 0x6c, 0x64, 0x49, 0x6e, 0x66, 0x6f, 0x53, 0x74, 0x61, 0x74, 0x73, 0x52, 0x05, 0x73, 0x74,
+	0x61, 0x74, 0x73, 0x22, 0x2c, 0x0a, 0x10, 0x42, 0x75, 0x69, 0x6c, 0x64, 0x49, 0x6e, 0x66, 0x6f,
+	0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x12, 0x18, 0x0a, 0x07, 0x76, 0x65, 0x72, 0x73, 0x69,
+	0x6f, 0x6e, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x76, 0x65, 0x72, 0x73, 0x69, 0x6f,
+	0x6e, 0x22, 0x85, 0x01, 0x0a, 0x0b, 0x42, 0x75, 0x69, 0x6c, 0x64, 0x49, 0x6e, 0x66, 0x6f, 0x4f,
+	0x53, 0x12, 0x12, 0x0a, 0x04, 0x74, 0x79, 0x70, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52,
+	0x04, 0x74, 0x79, 0x70, 0x65, 0x12, 0x1a, 0x0a, 0x08, 0x70, 0x6c, 0x61, 0x74, 0x66, 0x6f, 0x72,
+	0x6d, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x70, 0x6c, 0x61, 0x74, 0x66, 0x6f, 0x72,
+	0x6d, 0x12, 0x12, 0x0a, 0x04, 0x61, 0x72, 0x63, 0x68, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52,
+	0x04, 0x61, 0x72, 0x63, 0x68, 0x12, 0x18, 0x0a, 0x07, 0x76, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e,
+	0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x76, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x12,
+	0x18, 0x0a, 0x07, 0x72, 0x65, 0x6c, 0x65, 0x61, 0x73, 0x65, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09,
+	0x52, 0x07, 0x72, 0x65, 0x6c, 0x65, 0x61, 0x73, 0x65, 0x22, 0xea, 0x01, 0x0a, 0x0e, 0x42, 0x75,
+	0x69, 0x6c, 0x64, 0x49, 0x6e, 0x66, 0x6f, 0x53, 0x74, 0x61, 0x74, 0x73, 0x12, 0x1c, 0x0a, 0x09,
+	0x74, 0x6f, 0x74, 0x61, 0x6c, 0x41, 0x70, 0x69, 0x73, 0x18, 0x01, 0x20, 0x01, 0x28, 0x05, 0x52,
+	0x09, 0x74, 0x6f, 0x74, 0x61, 0x6c, 0x41, 0x70, 0x69, 0x73, 0x12, 0x28, 0x0a, 0x0f, 0x74, 0x6f,
+	0x74, 0x61, 0x6c, 0x4f, 0x70, 0x65, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x18, 0x02, 0x20,
+	0x01, 0x28, 0x05, 0x52, 0x0f, 0x74, 0x6f, 0x74, 0x61, 0x6c, 0x4f, 0x70, 0x65, 0x72, 0x61, 0x74,
+	0x69, 0x6f, 0x6e, 0x73, 0x12, 0x24, 0x0a, 0x0d, 0x74, 0x6f, 0x74, 0x61, 0x6c, 0x57, 0x65, 0x62,
+	0x68, 0x6f, 0x6f, 0x6b, 0x73, 0x18, 0x03, 0x20, 0x01, 0x28, 0x05, 0x52, 0x0d, 0x74, 0x6f, 0x74,
+	0x61, 0x6c, 0x57, 0x65, 0x62, 0x68, 0x6f, 0x6f, 0x6b, 0x73, 0x12, 0x3c, 0x0a, 0x19, 0x68, 0x61,
+	0x73, 0x41, 0x75, 0x74, 0x68, 0x65, 0x6e, 0x74, 0x69, 0x63, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x50,
+	0x72, 0x6f, 0x76, 0x69, 0x64, 0x65, 0x72, 0x18, 0x04, 0x20, 0x01, 0x28, 0x08, 0x52, 0x19, 0x68,
+	0x61, 0x73, 0x41, 0x75, 0x74, 0x68, 0x65, 0x6e, 0x74, 0x69, 0x63, 0x61, 0x74, 0x69, 0x6f, 0x6e,
+	0x50, 0x72, 0x6f, 0x76, 0x69, 0x64, 0x65, 0x72, 0x12, 0x2c, 0x0a, 0x11, 0x68, 0x61, 0x73, 0x55,
+	0x70, 0x6c, 0x6f, 0x61, 0x64, 0x50, 0x72, 0x6f, 0x76, 0x69, 0x64, 0x65, 0x72, 0x18, 0x05, 0x20,
+	0x01, 0x28, 0x08, 0x52, 0x11, 0x68, 0x61, 0x73, 0x55, 0x70, 0x6c, 0x6f, 0x61, 0x64, 0x50, 0x72,
+	0x6f, 0x76, 0x69, 0x64, 0x65, 0x72, 0x2a, 0x4d, 0x0a, 0x08, 0x4c, 0x6f, 0x67, 0x4c, 0x65, 0x76,
+	0x65, 0x6c, 0x12, 0x09, 0x0a, 0x05, 0x44, 0x45, 0x42, 0x55, 0x47, 0x10, 0x00, 0x12, 0x08, 0x0a,
+	0x04, 0x49, 0x4e, 0x46, 0x4f, 0x10, 0x01, 0x12, 0x09, 0x0a, 0x05, 0x45, 0x52, 0x52, 0x4f, 0x52,
+	0x10, 0x02, 0x12, 0x0b, 0x0a, 0x07, 0x57, 0x41, 0x52, 0x4e, 0x49, 0x4e, 0x47, 0x10, 0x03, 0x12,
+	0x09, 0x0a, 0x05, 0x50, 0x41, 0x4e, 0x49, 0x43, 0x10, 0x04, 0x12, 0x09, 0x0a, 0x05, 0x46, 0x41,
+	0x54, 0x41, 0x4c, 0x10, 0x05, 0x2a, 0x57, 0x0a, 0x10, 0x41, 0x75, 0x74, 0x68, 0x50, 0x72, 0x6f,
+	0x76, 0x69, 0x64, 0x65, 0x72, 0x4b, 0x69, 0x6e, 0x64, 0x12, 0x16, 0x0a, 0x12, 0x41, 0x75, 0x74,
+	0x68, 0x50, 0x72, 0x6f, 0x76, 0x69, 0x64, 0x65, 0x72, 0x47, 0x69, 0x74, 0x68, 0x75, 0x62, 0x10,
+	0x00, 0x12, 0x14, 0x0a, 0x10, 0x41, 0x75, 0x74, 0x68, 0x50, 0x72, 0x6f, 0x76, 0x69, 0x64, 0x65,
+	0x72, 0x4f, 0x49, 0x44, 0x43, 0x10, 0x01, 0x12, 0x15, 0x0a, 0x11, 0x41, 0x75, 0x74, 0x68, 0x50,
+	0x72, 0x6f, 0x76, 0x69, 0x64, 0x65, 0x72, 0x41, 0x75, 0x74, 0x68, 0x30, 0x10, 0x02, 0x2a, 0x41,
+	0x0a, 0x18, 0x4f, 0x70, 0x65, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x45, 0x78, 0x65, 0x63, 0x75,
+	0x74, 0x69, 0x6f, 0x6e, 0x45, 0x6e, 0x67, 0x69, 0x6e, 0x65, 0x12, 0x12, 0x0a, 0x0e, 0x45, 0x4e,
+	0x47, 0x49, 0x4e, 0x45, 0x5f, 0x47, 0x52, 0x41, 0x50, 0x48, 0x51, 0x4c, 0x10, 0x00, 0x12, 0x11,
+	0x0a, 0x0d, 0x45, 0x4e, 0x47, 0x49, 0x4e, 0x45, 0x5f, 0x4e, 0x4f, 0x44, 0x45, 0x4a, 0x53, 0x10,
+	0x01, 0x2a, 0x44, 0x0a, 0x1d, 0x50, 0x6f, 0x73, 0x74, 0x52, 0x65, 0x73, 0x6f, 0x6c, 0x76, 0x65,
+	0x54, 0x72, 0x61, 0x6e, 0x73, 0x66, 0x6f, 0x72, 0x6d, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x4b, 0x69,
+	0x6e, 0x64, 0x12, 0x23, 0x0a, 0x1f, 0x47, 0x45, 0x54, 0x5f, 0x50, 0x4f, 0x53, 0x54, 0x5f, 0x52,
+	0x45, 0x53, 0x4f, 0x4c, 0x56, 0x45, 0x5f, 0x54, 0x52, 0x41, 0x4e, 0x53, 0x46, 0x4f, 0x52, 0x4d,
+	0x41, 0x54, 0x49, 0x4f, 0x4e, 0x10, 0x00, 0x2a, 0x47, 0x0a, 0x12, 0x49, 0x6e, 0x6a, 0x65, 0x63,
+	0x74, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x4b, 0x69, 0x6e, 0x64, 0x12, 0x08, 0x0a,
+	0x04, 0x55, 0x55, 0x49, 0x44, 0x10, 0x00, 0x12, 0x0d, 0x0a, 0x09, 0x44, 0x41, 0x54, 0x45, 0x5f,
+	0x54, 0x49, 0x4d, 0x45, 0x10, 0x01, 0x12, 0x18, 0x0a, 0x14, 0x45, 0x4e, 0x56, 0x49, 0x52, 0x4f,
+	0x4e, 0x4d, 0x45, 0x4e, 0x54, 0x5f, 0x56, 0x41, 0x52, 0x49, 0x41, 0x42, 0x4c, 0x45, 0x10, 0x02,
+	0x2a, 0xbc, 0x02, 0x0a, 0x09, 0x43, 0x6c, 0x61, 0x69, 0x6d, 0x54, 0x79, 0x70, 0x65, 0x12, 0x0a,
+	0x0a, 0x06, 0x49, 0x53, 0x53, 0x55, 0x45, 0x52, 0x10, 0x00, 0x12, 0x0c, 0x0a, 0x08, 0x50, 0x52,
+	0x4f, 0x56, 0x49, 0x44, 0x45, 0x52, 0x10, 0x00, 0x12, 0x0b, 0x0a, 0x07, 0x53, 0x55, 0x42, 0x4a,
+	0x45, 0x43, 0x54, 0x10, 0x01, 0x12, 0x0a, 0x0a, 0x06, 0x55, 0x53, 0x45, 0x52, 0x49, 0x44, 0x10,
+	0x01, 0x12, 0x08, 0x0a, 0x04, 0x4e, 0x41, 0x4d, 0x45, 0x10, 0x02, 0x12, 0x0e, 0x0a, 0x0a, 0x47,
+	0x49, 0x56, 0x45, 0x4e, 0x5f, 0x4e, 0x41, 0x4d, 0x45, 0x10, 0x03, 0x12, 0x0f, 0x0a, 0x0b, 0x46,
+	0x41, 0x4d, 0x49, 0x4c, 0x59, 0x5f, 0x4e, 0x41, 0x4d, 0x45, 0x10, 0x04, 0x12, 0x0f, 0x0a, 0x0b,
+	0x4d, 0x49, 0x44, 0x44, 0x4c, 0x45, 0x5f, 0x4e, 0x41, 0x4d, 0x45, 0x10, 0x05, 0x12, 0x0c, 0x0a,
+	0x08, 0x4e, 0x49, 0x43, 0x4b, 0x4e, 0x41, 0x4d, 0x45, 0x10, 0x06, 0x12, 0x16, 0x0a, 0x12, 0x50,
+	0x52, 0x45, 0x46, 0x45, 0x52, 0x52, 0x45, 0x44, 0x5f, 0x55, 0x53, 0x45, 0x52, 0x4e, 0x41, 0x4d,
+	0x45, 0x10, 0x07, 0x12, 0x0b, 0x0a, 0x07, 0x50, 0x52, 0x4f, 0x46, 0x49, 0x4c, 0x45, 0x10, 0x08,
+	0x12, 0x0b, 0x0a, 0x07, 0x50, 0x49, 0x43, 0x54, 0x55, 0x52, 0x45, 0x10, 0x09, 0x12, 0x0b, 0x0a,
+	0x07, 0x57, 0x45, 0x42, 0x53, 0x49, 0x54, 0x45, 0x10, 0x0a, 0x12, 0x09, 0x0a, 0x05, 0x45, 0x4d,
+	0x41, 0x49, 0x4c, 0x10, 0x0b, 0x12, 0x12, 0x0a, 0x0e, 0x45, 0x4d, 0x41, 0x49, 0x4c, 0x5f, 0x56,
+	0x45, 0x52, 0x49, 0x46, 0x49, 0x45, 0x44, 0x10, 0x0c, 0x12, 0x0a, 0x0a, 0x06, 0x47, 0x45, 0x4e,
+	0x44, 0x45, 0x52, 0x10, 0x0d, 0x12, 0x0e, 0x0a, 0x0a, 0x42, 0x49, 0x52, 0x54, 0x48, 0x5f, 0x44,
+	0x41, 0x54, 0x45, 0x10, 0x0e, 0x12, 0x0d, 0x0a, 0x09, 0x5a, 0x4f, 0x4e, 0x45, 0x5f, 0x49, 0x4e,
+	0x46, 0x4f, 0x10, 0x0f, 0x12, 0x0a, 0x0a, 0x06, 0x4c, 0x4f, 0x43, 0x41, 0x4c, 0x45, 0x10, 0x10,
+	0x12, 0x0c, 0x0a, 0x08, 0x4c, 0x4f, 0x43, 0x41, 0x54, 0x49, 0x4f, 0x4e, 0x10, 0x11, 0x12, 0x0b,
+	0x0a, 0x06, 0x43, 0x55, 0x53, 0x54, 0x4f, 0x4d, 0x10, 0xe7, 0x07, 0x1a, 0x02, 0x10, 0x01, 0x2a,
+	0x41, 0x0a, 0x09, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x54, 0x79, 0x70, 0x65, 0x12, 0x0a, 0x0a, 0x06,
+	0x53, 0x54, 0x52, 0x49, 0x4e, 0x47, 0x10, 0x00, 0x12, 0x07, 0x0a, 0x03, 0x49, 0x4e, 0x54, 0x10,
+	0x01, 0x12, 0x09, 0x0a, 0x05, 0x46, 0x4c, 0x4f, 0x41, 0x54, 0x10, 0x02, 0x12, 0x0b, 0x0a, 0x07,
+	0x42, 0x4f, 0x4f, 0x4c, 0x45, 0x41, 0x4e, 0x10, 0x03, 0x12, 0x07, 0x0a, 0x03, 0x41, 0x4e, 0x59,
+	0x10, 0x04, 0x2a, 0x3a, 0x0a, 0x0d, 0x4f, 0x70, 0x65, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x54,
+	0x79, 0x70, 0x65, 0x12, 0x09, 0x0a, 0x05, 0x51, 0x55, 0x45, 0x52, 0x59, 0x10, 0x00, 0x12, 0x0c,
+	0x0a, 0x08, 0x4d, 0x55, 0x54, 0x41, 0x54, 0x49, 0x4f, 0x4e, 0x10, 0x01, 0x12, 0x10, 0x0a, 0x0c,
+	0x53, 0x55, 0x42, 0x53, 0x43, 0x52, 0x49, 0x50, 0x54, 0x49, 0x4f, 0x4e, 0x10, 0x02, 0x2a, 0x82,
+	0x01, 0x0a, 0x0e, 0x44, 0x61, 0x74, 0x61, 0x53, 0x6f, 0x75, 0x72, 0x63, 0x65, 0x4b, 0x69, 0x6e,
+	0x64, 0x12, 0x0a, 0x0a, 0x06, 0x53, 0x54, 0x41, 0x54, 0x49, 0x43, 0x10, 0x00, 0x12, 0x08, 0x0a,
+	0x04, 0x52, 0x45, 0x53, 0x54, 0x10, 0x01, 0x12, 0x0b, 0x0a, 0x07, 0x47, 0x52, 0x41, 0x50, 0x48,
+	0x51, 0x4c, 0x10, 0x02, 0x12, 0x0e, 0x0a, 0x0a, 0x50, 0x4f, 0x53, 0x54, 0x47, 0x52, 0x45, 0x53,
+	0x51, 0x4c, 0x10, 0x03, 0x12, 0x09, 0x0a, 0x05, 0x4d, 0x59, 0x53, 0x51, 0x4c, 0x10, 0x04, 0x12,
+	0x0d, 0x0a, 0x09, 0x53, 0x51, 0x4c, 0x53, 0x45, 0x52, 0x56, 0x45, 0x52, 0x10, 0x05, 0x12, 0x0b,
+	0x0a, 0x07, 0x4d, 0x4f, 0x4e, 0x47, 0x4f, 0x44, 0x42, 0x10, 0x06, 0x12, 0x0a, 0x0a, 0x06, 0x53,
+	0x51, 0x4c, 0x49, 0x54, 0x45, 0x10, 0x07, 0x12, 0x0a, 0x0a, 0x06, 0x50, 0x52, 0x49, 0x53, 0x4d,
+	0x41, 0x10, 0x08, 0x2a, 0x71, 0x0a, 0x1a, 0x55, 0x70, 0x73, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x41,
+	0x75, 0x74, 0x68, 0x65, 0x6e, 0x74, 0x69, 0x63, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x4b, 0x69, 0x6e,
+	0x64, 0x12, 0x1d, 0x0a, 0x19, 0x55, 0x70, 0x73, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x41, 0x75, 0x74,
+	0x68, 0x65, 0x6e, 0x74, 0x69, 0x63, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x4a, 0x57, 0x54, 0x10, 0x00,
+	0x12, 0x34, 0x0a, 0x30, 0x55, 0x70, 0x73, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x41, 0x75, 0x74, 0x68,
+	0x65, 0x6e, 0x74, 0x69, 0x63, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x4a, 0x57, 0x54, 0x57, 0x69, 0x74,
+	0x68, 0x41, 0x63, 0x63, 0x65, 0x73, 0x73, 0x54, 0x6f, 0x6b, 0x65, 0x6e, 0x45, 0x78, 0x63, 0x68,
+	0x61, 0x6e, 0x67, 0x65, 0x10, 0x01, 0x2a, 0x27, 0x0a, 0x0d, 0x53, 0x69, 0x67, 0x6e, 0x69, 0x6e,
+	0x67, 0x4d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x12, 0x16, 0x0a, 0x12, 0x53, 0x69, 0x67, 0x6e, 0x69,
+	0x6e, 0x67, 0x4d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x48, 0x53, 0x32, 0x35, 0x36, 0x10, 0x00, 0x2a,
+	0x41, 0x0a, 0x0a, 0x48, 0x54, 0x54, 0x50, 0x4d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x12, 0x07, 0x0a,
+	0x03, 0x47, 0x45, 0x54, 0x10, 0x00, 0x12, 0x08, 0x0a, 0x04, 0x50, 0x4f, 0x53, 0x54, 0x10, 0x01,
+	0x12, 0x07, 0x0a, 0x03, 0x50, 0x55, 0x54, 0x10, 0x02, 0x12, 0x0a, 0x0a, 0x06, 0x44, 0x45, 0x4c,
+	0x45, 0x54, 0x45, 0x10, 0x03, 0x12, 0x0b, 0x0a, 0x07, 0x4f, 0x50, 0x54, 0x49, 0x4f, 0x4e, 0x53,
+	0x10, 0x04, 0x2a, 0x36, 0x0a, 0x0e, 0x41, 0x72, 0x67, 0x75, 0x6d, 0x65, 0x6e, 0x74, 0x53, 0x6f,
+	0x75, 0x72, 0x63, 0x65, 0x12, 0x10, 0x0a, 0x0c, 0x4f, 0x42, 0x4a, 0x45, 0x43, 0x54, 0x5f, 0x46,
+	0x49, 0x45, 0x4c, 0x44, 0x10, 0x00, 0x12, 0x12, 0x0a, 0x0e, 0x46, 0x49, 0x45, 0x4c, 0x44, 0x5f,
+	0x41, 0x52, 0x47, 0x55, 0x4d, 0x45, 0x4e, 0x54, 0x10, 0x01, 0x2a, 0x82, 0x01, 0x0a, 0x1b, 0x41,
+	0x72, 0x67, 0x75, 0x6d, 0x65, 0x6e, 0x74, 0x52, 0x65, 0x6e, 0x64, 0x65, 0x72, 0x43, 0x6f, 0x6e,
+	0x66, 0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x1b, 0x0a, 0x17, 0x52, 0x45,
+	0x4e, 0x44, 0x45, 0x52, 0x5f, 0x41, 0x52, 0x47, 0x55, 0x4d, 0x45, 0x4e, 0x54, 0x5f, 0x44, 0x45,
+	0x46, 0x41, 0x55, 0x4c, 0x54, 0x10, 0x00, 0x12, 0x24, 0x0a, 0x20, 0x52, 0x45, 0x4e, 0x44, 0x45,
+	0x52, 0x5f, 0x41, 0x52, 0x47, 0x55, 0x4d, 0x45, 0x4e, 0x54, 0x5f, 0x41, 0x53, 0x5f, 0x47, 0x52,
+	0x41, 0x50, 0x48, 0x51, 0x4c, 0x5f, 0x56, 0x41, 0x4c, 0x55, 0x45, 0x10, 0x01, 0x12, 0x20, 0x0a,
+	0x1c, 0x52, 0x45, 0x4e, 0x44, 0x45, 0x52, 0x5f, 0x41, 0x52, 0x47, 0x55, 0x4d, 0x45, 0x4e, 0x54,
+	0x5f, 0x41, 0x53, 0x5f, 0x41, 0x52, 0x52, 0x41, 0x59, 0x5f, 0x43, 0x53, 0x56, 0x10, 0x02, 0x2a,
+	0x26, 0x0a, 0x13, 0x57, 0x65, 0x62, 0x68, 0x6f, 0x6f, 0x6b, 0x56, 0x65, 0x72, 0x69, 0x66, 0x69,
+	0x65, 0x72, 0x4b, 0x69, 0x6e, 0x64, 0x12, 0x0f, 0x0a, 0x0b, 0x48, 0x4d, 0x41, 0x43, 0x5f, 0x53,
+	0x48, 0x41, 0x32, 0x35, 0x36, 0x10, 0x00, 0x2a, 0x86, 0x01, 0x0a, 0x19, 0x43, 0x6f, 0x6e, 0x66,
+	0x69, 0x67, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c,
+	0x65, 0x4b, 0x69, 0x6e, 0x64, 0x12, 0x21, 0x0a, 0x1d, 0x53, 0x54, 0x41, 0x54, 0x49, 0x43, 0x5f,
+	0x43, 0x4f, 0x4e, 0x46, 0x49, 0x47, 0x55, 0x52, 0x41, 0x54, 0x49, 0x4f, 0x4e, 0x5f, 0x56, 0x41,
+	0x52, 0x49, 0x41, 0x42, 0x4c, 0x45, 0x10, 0x00, 0x12, 0x1e, 0x0a, 0x1a, 0x45, 0x4e, 0x56, 0x5f,
+	0x43, 0x4f, 0x4e, 0x46, 0x49, 0x47, 0x55, 0x52, 0x41, 0x54, 0x49, 0x4f, 0x4e, 0x5f, 0x56, 0x41,
+	0x52, 0x49, 0x41, 0x42, 0x4c, 0x45, 0x10, 0x01, 0x12, 0x26, 0x0a, 0x22, 0x50, 0x4c, 0x41, 0x43,
+	0x45, 0x48, 0x4f, 0x4c, 0x44, 0x45, 0x52, 0x5f, 0x43, 0x4f, 0x4e, 0x46, 0x49, 0x47, 0x55, 0x52,
+	0x41, 0x54, 0x49, 0x4f, 0x4e, 0x5f, 0x56, 0x41, 0x52, 0x49, 0x41, 0x42, 0x4c, 0x45, 0x10, 0x02,
+	0x42, 0x2d, 0x5a, 0x2b, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x77,
+	0x75, 0x6e, 0x64, 0x65, 0x72, 0x67, 0x72, 0x61, 0x70, 0x68, 0x2f, 0x77, 0x75, 0x6e, 0x64, 0x65,
+	0x72, 0x67, 0x72, 0x61, 0x70, 0x68, 0x2f, 0x70, 0x6b, 0x67, 0x2f, 0x77, 0x67, 0x70, 0x62, 0x62,
+	0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
@@ -6801,7 +6901,7 @@ func file_wundernode_config_proto_rawDescGZIP() []byte {
 }
 
 var file_wundernode_config_proto_enumTypes = make([]protoimpl.EnumInfo, 16)
-var file_wundernode_config_proto_msgTypes = make([]protoimpl.MessageInfo, 70)
+var file_wundernode_config_proto_msgTypes = make([]protoimpl.MessageInfo, 71)
 var file_wundernode_config_proto_goTypes = []interface{}{
 	(LogLevel)(0),                                            // 0: wgpb.LogLevel
 	(AuthProviderKind)(0),                                    // 1: wgpb.AuthProviderKind
@@ -6877,44 +6977,45 @@ var file_wundernode_config_proto_goTypes = []interface{}{
 	(*NodeLogging)(nil),                                      // 71: wgpb.NodeLogging
 	(*PrometheusOptions)(nil),                                // 72: wgpb.PrometheusOptions
 	(*NodeOptions)(nil),                                      // 73: wgpb.NodeOptions
-	(*ServerLogging)(nil),                                    // 74: wgpb.ServerLogging
-	(*ServerOptions)(nil),                                    // 75: wgpb.ServerOptions
-	(*WebhookConfiguration)(nil),                             // 76: wgpb.WebhookConfiguration
-	(*WebhookVerifier)(nil),                                  // 77: wgpb.WebhookVerifier
-	(*CorsConfiguration)(nil),                                // 78: wgpb.CorsConfiguration
-	(*ConfigurationVariable)(nil),                            // 79: wgpb.ConfigurationVariable
-	(*BuildInfo)(nil),                                        // 80: wgpb.BuildInfo
-	(*BuildInfoVersion)(nil),                                 // 81: wgpb.BuildInfoVersion
-	(*BuildInfoOS)(nil),                                      // 82: wgpb.BuildInfoOS
-	(*BuildInfoStats)(nil),                                   // 83: wgpb.BuildInfoStats
-	nil,                                                      // 84: wgpb.FetchConfiguration.HeaderEntry
-	nil,                                                      // 85: wgpb.S3UploadConfiguration.UploadProfilesEntry
+	(*TelemetryOptions)(nil),                                 // 74: wgpb.TelemetryOptions
+	(*ServerLogging)(nil),                                    // 75: wgpb.ServerLogging
+	(*ServerOptions)(nil),                                    // 76: wgpb.ServerOptions
+	(*WebhookConfiguration)(nil),                             // 77: wgpb.WebhookConfiguration
+	(*WebhookVerifier)(nil),                                  // 78: wgpb.WebhookVerifier
+	(*CorsConfiguration)(nil),                                // 79: wgpb.CorsConfiguration
+	(*ConfigurationVariable)(nil),                            // 80: wgpb.ConfigurationVariable
+	(*BuildInfo)(nil),                                        // 81: wgpb.BuildInfo
+	(*BuildInfoVersion)(nil),                                 // 82: wgpb.BuildInfoVersion
+	(*BuildInfoOS)(nil),                                      // 83: wgpb.BuildInfoOS
+	(*BuildInfoStats)(nil),                                   // 84: wgpb.BuildInfoStats
+	nil,                                                      // 85: wgpb.FetchConfiguration.HeaderEntry
+	nil,                                                      // 86: wgpb.S3UploadConfiguration.UploadProfilesEntry
 }
 var file_wundernode_config_proto_depIdxs = []int32{
 	20,  // 0: wgpb.ApiAuthenticationConfig.cookieBased:type_name -> wgpb.CookieBasedAuthentication
 	19,  // 1: wgpb.ApiAuthenticationConfig.hooks:type_name -> wgpb.ApiAuthenticationHooks
 	17,  // 2: wgpb.ApiAuthenticationConfig.jwksBased:type_name -> wgpb.JwksBasedAuthentication
 	18,  // 3: wgpb.JwksBasedAuthentication.providers:type_name -> wgpb.JwksAuthProvider
-	79,  // 4: wgpb.JwksAuthProvider.jwksUrl:type_name -> wgpb.ConfigurationVariable
-	79,  // 5: wgpb.JwksAuthProvider.jwksJson:type_name -> wgpb.ConfigurationVariable
-	79,  // 6: wgpb.JwksAuthProvider.userInfoEndpoint:type_name -> wgpb.ConfigurationVariable
+	80,  // 4: wgpb.JwksAuthProvider.jwksUrl:type_name -> wgpb.ConfigurationVariable
+	80,  // 5: wgpb.JwksAuthProvider.jwksJson:type_name -> wgpb.ConfigurationVariable
+	80,  // 6: wgpb.JwksAuthProvider.userInfoEndpoint:type_name -> wgpb.ConfigurationVariable
 	21,  // 7: wgpb.CookieBasedAuthentication.providers:type_name -> wgpb.AuthProvider
-	79,  // 8: wgpb.CookieBasedAuthentication.authorizedRedirectUris:type_name -> wgpb.ConfigurationVariable
-	79,  // 9: wgpb.CookieBasedAuthentication.authorizedRedirectUriRegexes:type_name -> wgpb.ConfigurationVariable
-	79,  // 10: wgpb.CookieBasedAuthentication.hashKey:type_name -> wgpb.ConfigurationVariable
-	79,  // 11: wgpb.CookieBasedAuthentication.blockKey:type_name -> wgpb.ConfigurationVariable
-	79,  // 12: wgpb.CookieBasedAuthentication.csrfSecret:type_name -> wgpb.ConfigurationVariable
-	79,  // 13: wgpb.CookieBasedAuthentication.timeoutSeconds:type_name -> wgpb.ConfigurationVariable
+	80,  // 8: wgpb.CookieBasedAuthentication.authorizedRedirectUris:type_name -> wgpb.ConfigurationVariable
+	80,  // 9: wgpb.CookieBasedAuthentication.authorizedRedirectUriRegexes:type_name -> wgpb.ConfigurationVariable
+	80,  // 10: wgpb.CookieBasedAuthentication.hashKey:type_name -> wgpb.ConfigurationVariable
+	80,  // 11: wgpb.CookieBasedAuthentication.blockKey:type_name -> wgpb.ConfigurationVariable
+	80,  // 12: wgpb.CookieBasedAuthentication.csrfSecret:type_name -> wgpb.ConfigurationVariable
+	80,  // 13: wgpb.CookieBasedAuthentication.timeoutSeconds:type_name -> wgpb.ConfigurationVariable
 	1,   // 14: wgpb.AuthProvider.kind:type_name -> wgpb.AuthProviderKind
 	22,  // 15: wgpb.AuthProvider.githubConfig:type_name -> wgpb.GithubAuthProviderConfig
 	24,  // 16: wgpb.AuthProvider.oidcConfig:type_name -> wgpb.OpenIDConnectAuthProviderConfig
-	79,  // 17: wgpb.GithubAuthProviderConfig.clientId:type_name -> wgpb.ConfigurationVariable
-	79,  // 18: wgpb.GithubAuthProviderConfig.clientSecret:type_name -> wgpb.ConfigurationVariable
-	79,  // 19: wgpb.OpenIDConnectQueryParameter.name:type_name -> wgpb.ConfigurationVariable
-	79,  // 20: wgpb.OpenIDConnectQueryParameter.value:type_name -> wgpb.ConfigurationVariable
-	79,  // 21: wgpb.OpenIDConnectAuthProviderConfig.issuer:type_name -> wgpb.ConfigurationVariable
-	79,  // 22: wgpb.OpenIDConnectAuthProviderConfig.clientId:type_name -> wgpb.ConfigurationVariable
-	79,  // 23: wgpb.OpenIDConnectAuthProviderConfig.clientSecret:type_name -> wgpb.ConfigurationVariable
+	80,  // 17: wgpb.GithubAuthProviderConfig.clientId:type_name -> wgpb.ConfigurationVariable
+	80,  // 18: wgpb.GithubAuthProviderConfig.clientSecret:type_name -> wgpb.ConfigurationVariable
+	80,  // 19: wgpb.OpenIDConnectQueryParameter.name:type_name -> wgpb.ConfigurationVariable
+	80,  // 20: wgpb.OpenIDConnectQueryParameter.value:type_name -> wgpb.ConfigurationVariable
+	80,  // 21: wgpb.OpenIDConnectAuthProviderConfig.issuer:type_name -> wgpb.ConfigurationVariable
+	80,  // 22: wgpb.OpenIDConnectAuthProviderConfig.clientId:type_name -> wgpb.ConfigurationVariable
+	80,  // 23: wgpb.OpenIDConnectAuthProviderConfig.clientSecret:type_name -> wgpb.ConfigurationVariable
 	23,  // 24: wgpb.OpenIDConnectAuthProviderConfig.queryParameters:type_name -> wgpb.OpenIDConnectQueryParameter
 	7,   // 25: wgpb.Operation.operationType:type_name -> wgpb.OperationType
 	39,  // 26: wgpb.Operation.cacheConfig:type_name -> wgpb.OperationCacheConfig
@@ -6954,87 +7055,92 @@ var file_wundernode_config_proto_depIdxs = []int32{
 	47,  // 60: wgpb.DataSourceCustom_GraphQL.federation:type_name -> wgpb.GraphQLFederationConfiguration
 	30,  // 61: wgpb.DataSourceCustom_GraphQL.hooksConfiguration:type_name -> wgpb.GraphQLDataSourceHooksConfiguration
 	61,  // 62: wgpb.DataSourceCustom_GraphQL.customScalarTypeFields:type_name -> wgpb.SingleTypeField
-	79,  // 63: wgpb.DataSourceCustom_Database.databaseURL:type_name -> wgpb.ConfigurationVariable
+	80,  // 63: wgpb.DataSourceCustom_Database.databaseURL:type_name -> wgpb.ConfigurationVariable
 	61,  // 64: wgpb.DataSourceCustom_Database.jsonTypeFields:type_name -> wgpb.SingleTypeField
-	79,  // 65: wgpb.DataSourceCustom_Static.data:type_name -> wgpb.ConfigurationVariable
-	79,  // 66: wgpb.GraphQLSubscriptionConfiguration.url:type_name -> wgpb.ConfigurationVariable
-	79,  // 67: wgpb.FetchConfiguration.url:type_name -> wgpb.ConfigurationVariable
+	80,  // 65: wgpb.DataSourceCustom_Static.data:type_name -> wgpb.ConfigurationVariable
+	80,  // 66: wgpb.GraphQLSubscriptionConfiguration.url:type_name -> wgpb.ConfigurationVariable
+	80,  // 67: wgpb.FetchConfiguration.url:type_name -> wgpb.ConfigurationVariable
 	11,  // 68: wgpb.FetchConfiguration.method:type_name -> wgpb.HTTPMethod
-	84,  // 69: wgpb.FetchConfiguration.header:type_name -> wgpb.FetchConfiguration.HeaderEntry
-	79,  // 70: wgpb.FetchConfiguration.body:type_name -> wgpb.ConfigurationVariable
+	85,  // 69: wgpb.FetchConfiguration.header:type_name -> wgpb.FetchConfiguration.HeaderEntry
+	80,  // 70: wgpb.FetchConfiguration.body:type_name -> wgpb.ConfigurationVariable
 	56,  // 71: wgpb.FetchConfiguration.query:type_name -> wgpb.URLQueryConfiguration
 	52,  // 72: wgpb.FetchConfiguration.upstreamAuthentication:type_name -> wgpb.UpstreamAuthentication
 	51,  // 73: wgpb.FetchConfiguration.mTLS:type_name -> wgpb.MTLSConfiguration
-	79,  // 74: wgpb.FetchConfiguration.baseUrl:type_name -> wgpb.ConfigurationVariable
-	79,  // 75: wgpb.FetchConfiguration.path:type_name -> wgpb.ConfigurationVariable
-	79,  // 76: wgpb.FetchConfiguration.httpProxyUrl:type_name -> wgpb.ConfigurationVariable
-	79,  // 77: wgpb.MTLSConfiguration.key:type_name -> wgpb.ConfigurationVariable
-	79,  // 78: wgpb.MTLSConfiguration.cert:type_name -> wgpb.ConfigurationVariable
+	80,  // 74: wgpb.FetchConfiguration.baseUrl:type_name -> wgpb.ConfigurationVariable
+	80,  // 75: wgpb.FetchConfiguration.path:type_name -> wgpb.ConfigurationVariable
+	80,  // 76: wgpb.FetchConfiguration.httpProxyUrl:type_name -> wgpb.ConfigurationVariable
+	80,  // 77: wgpb.MTLSConfiguration.key:type_name -> wgpb.ConfigurationVariable
+	80,  // 78: wgpb.MTLSConfiguration.cert:type_name -> wgpb.ConfigurationVariable
 	9,   // 79: wgpb.UpstreamAuthentication.kind:type_name -> wgpb.UpstreamAuthenticationKind
 	53,  // 80: wgpb.UpstreamAuthentication.jwtConfig:type_name -> wgpb.JwtUpstreamAuthenticationConfig
 	54,  // 81: wgpb.UpstreamAuthentication.jwtWithAccessTokenExchangeConfig:type_name -> wgpb.JwtUpstreamAuthenticationWithAccessTokenExchange
-	79,  // 82: wgpb.JwtUpstreamAuthenticationConfig.secret:type_name -> wgpb.ConfigurationVariable
+	80,  // 82: wgpb.JwtUpstreamAuthenticationConfig.secret:type_name -> wgpb.ConfigurationVariable
 	10,  // 83: wgpb.JwtUpstreamAuthenticationConfig.signingMethod:type_name -> wgpb.SigningMethod
-	79,  // 84: wgpb.JwtUpstreamAuthenticationWithAccessTokenExchange.secret:type_name -> wgpb.ConfigurationVariable
+	80,  // 84: wgpb.JwtUpstreamAuthenticationWithAccessTokenExchange.secret:type_name -> wgpb.ConfigurationVariable
 	10,  // 85: wgpb.JwtUpstreamAuthenticationWithAccessTokenExchange.signingMethod:type_name -> wgpb.SigningMethod
-	79,  // 86: wgpb.JwtUpstreamAuthenticationWithAccessTokenExchange.accessTokenExchangeEndpoint:type_name -> wgpb.ConfigurationVariable
-	79,  // 87: wgpb.HTTPHeader.values:type_name -> wgpb.ConfigurationVariable
+	80,  // 86: wgpb.JwtUpstreamAuthenticationWithAccessTokenExchange.accessTokenExchangeEndpoint:type_name -> wgpb.ConfigurationVariable
+	80,  // 87: wgpb.HTTPHeader.values:type_name -> wgpb.ConfigurationVariable
 	62,  // 88: wgpb.FieldConfiguration.argumentsConfiguration:type_name -> wgpb.ArgumentConfiguration
 	12,  // 89: wgpb.ArgumentConfiguration.sourceType:type_name -> wgpb.ArgumentSource
 	13,  // 90: wgpb.ArgumentConfiguration.renderConfiguration:type_name -> wgpb.ArgumentRenderConfiguration
 	67,  // 91: wgpb.WunderGraphConfiguration.api:type_name -> wgpb.UserDefinedApi
 	64,  // 92: wgpb.S3UploadProfile.hooks:type_name -> wgpb.S3UploadProfileHooksConfiguration
-	79,  // 93: wgpb.S3UploadConfiguration.endpoint:type_name -> wgpb.ConfigurationVariable
-	79,  // 94: wgpb.S3UploadConfiguration.accessKeyID:type_name -> wgpb.ConfigurationVariable
-	79,  // 95: wgpb.S3UploadConfiguration.secretAccessKey:type_name -> wgpb.ConfigurationVariable
-	79,  // 96: wgpb.S3UploadConfiguration.bucketName:type_name -> wgpb.ConfigurationVariable
-	79,  // 97: wgpb.S3UploadConfiguration.bucketLocation:type_name -> wgpb.ConfigurationVariable
-	85,  // 98: wgpb.S3UploadConfiguration.uploadProfiles:type_name -> wgpb.S3UploadConfiguration.UploadProfilesEntry
+	80,  // 93: wgpb.S3UploadConfiguration.endpoint:type_name -> wgpb.ConfigurationVariable
+	80,  // 94: wgpb.S3UploadConfiguration.accessKeyID:type_name -> wgpb.ConfigurationVariable
+	80,  // 95: wgpb.S3UploadConfiguration.secretAccessKey:type_name -> wgpb.ConfigurationVariable
+	80,  // 96: wgpb.S3UploadConfiguration.bucketName:type_name -> wgpb.ConfigurationVariable
+	80,  // 97: wgpb.S3UploadConfiguration.bucketLocation:type_name -> wgpb.ConfigurationVariable
+	86,  // 98: wgpb.S3UploadConfiguration.uploadProfiles:type_name -> wgpb.S3UploadConfiguration.UploadProfilesEntry
 	40,  // 99: wgpb.UserDefinedApi.engineConfiguration:type_name -> wgpb.EngineConfiguration
 	25,  // 100: wgpb.UserDefinedApi.operations:type_name -> wgpb.Operation
-	78,  // 101: wgpb.UserDefinedApi.corsConfiguration:type_name -> wgpb.CorsConfiguration
+	79,  // 101: wgpb.UserDefinedApi.corsConfiguration:type_name -> wgpb.CorsConfiguration
 	16,  // 102: wgpb.UserDefinedApi.authenticationConfig:type_name -> wgpb.ApiAuthenticationConfig
 	66,  // 103: wgpb.UserDefinedApi.s3UploadConfiguration:type_name -> wgpb.S3UploadConfiguration
-	79,  // 104: wgpb.UserDefinedApi.allowedHostNames:type_name -> wgpb.ConfigurationVariable
-	76,  // 105: wgpb.UserDefinedApi.webhooks:type_name -> wgpb.WebhookConfiguration
-	75,  // 106: wgpb.UserDefinedApi.serverOptions:type_name -> wgpb.ServerOptions
+	80,  // 104: wgpb.UserDefinedApi.allowedHostNames:type_name -> wgpb.ConfigurationVariable
+	77,  // 105: wgpb.UserDefinedApi.webhooks:type_name -> wgpb.WebhookConfiguration
+	76,  // 106: wgpb.UserDefinedApi.serverOptions:type_name -> wgpb.ServerOptions
 	73,  // 107: wgpb.UserDefinedApi.nodeOptions:type_name -> wgpb.NodeOptions
 	68,  // 108: wgpb.UserDefinedApi.experimentalConfig:type_name -> wgpb.ExperimentalConfiguration
-	79,  // 109: wgpb.ListenerOptions.host:type_name -> wgpb.ConfigurationVariable
-	79,  // 110: wgpb.ListenerOptions.port:type_name -> wgpb.ConfigurationVariable
-	79,  // 111: wgpb.InternalListenerOptions.port:type_name -> wgpb.ConfigurationVariable
-	79,  // 112: wgpb.NodeLogging.level:type_name -> wgpb.ConfigurationVariable
-	79,  // 113: wgpb.PrometheusOptions.enabled:type_name -> wgpb.ConfigurationVariable
-	79,  // 114: wgpb.PrometheusOptions.port:type_name -> wgpb.ConfigurationVariable
-	79,  // 115: wgpb.NodeOptions.nodeUrl:type_name -> wgpb.ConfigurationVariable
-	79,  // 116: wgpb.NodeOptions.publicNodeUrl:type_name -> wgpb.ConfigurationVariable
+	80,  // 109: wgpb.ListenerOptions.host:type_name -> wgpb.ConfigurationVariable
+	80,  // 110: wgpb.ListenerOptions.port:type_name -> wgpb.ConfigurationVariable
+	80,  // 111: wgpb.InternalListenerOptions.port:type_name -> wgpb.ConfigurationVariable
+	80,  // 112: wgpb.NodeLogging.level:type_name -> wgpb.ConfigurationVariable
+	80,  // 113: wgpb.PrometheusOptions.enabled:type_name -> wgpb.ConfigurationVariable
+	80,  // 114: wgpb.PrometheusOptions.port:type_name -> wgpb.ConfigurationVariable
+	80,  // 115: wgpb.NodeOptions.nodeUrl:type_name -> wgpb.ConfigurationVariable
+	80,  // 116: wgpb.NodeOptions.publicNodeUrl:type_name -> wgpb.ConfigurationVariable
 	69,  // 117: wgpb.NodeOptions.listen:type_name -> wgpb.ListenerOptions
 	71,  // 118: wgpb.NodeOptions.logger:type_name -> wgpb.NodeLogging
 	70,  // 119: wgpb.NodeOptions.listenInternal:type_name -> wgpb.InternalListenerOptions
-	79,  // 120: wgpb.NodeOptions.nodeInternalUrl:type_name -> wgpb.ConfigurationVariable
-	79,  // 121: wgpb.NodeOptions.defaultHttpProxyUrl:type_name -> wgpb.ConfigurationVariable
-	72,  // 122: wgpb.NodeOptions.prometheus:type_name -> wgpb.PrometheusOptions
-	79,  // 123: wgpb.ServerLogging.level:type_name -> wgpb.ConfigurationVariable
-	79,  // 124: wgpb.ServerOptions.serverUrl:type_name -> wgpb.ConfigurationVariable
-	69,  // 125: wgpb.ServerOptions.listen:type_name -> wgpb.ListenerOptions
-	74,  // 126: wgpb.ServerOptions.logger:type_name -> wgpb.ServerLogging
-	77,  // 127: wgpb.WebhookConfiguration.verifier:type_name -> wgpb.WebhookVerifier
-	14,  // 128: wgpb.WebhookVerifier.kind:type_name -> wgpb.WebhookVerifierKind
-	79,  // 129: wgpb.WebhookVerifier.secret:type_name -> wgpb.ConfigurationVariable
-	79,  // 130: wgpb.CorsConfiguration.allowedOrigins:type_name -> wgpb.ConfigurationVariable
-	15,  // 131: wgpb.ConfigurationVariable.kind:type_name -> wgpb.ConfigurationVariableKind
-	81,  // 132: wgpb.BuildInfo.sdk:type_name -> wgpb.BuildInfoVersion
-	81,  // 133: wgpb.BuildInfo.wunderctl:type_name -> wgpb.BuildInfoVersion
-	81,  // 134: wgpb.BuildInfo.node:type_name -> wgpb.BuildInfoVersion
-	82,  // 135: wgpb.BuildInfo.os:type_name -> wgpb.BuildInfoOS
-	83,  // 136: wgpb.BuildInfo.stats:type_name -> wgpb.BuildInfoStats
-	57,  // 137: wgpb.FetchConfiguration.HeaderEntry.value:type_name -> wgpb.HTTPHeader
-	65,  // 138: wgpb.S3UploadConfiguration.UploadProfilesEntry.value:type_name -> wgpb.S3UploadProfile
-	139, // [139:139] is the sub-list for method output_type
-	139, // [139:139] is the sub-list for method input_type
-	139, // [139:139] is the sub-list for extension type_name
-	139, // [139:139] is the sub-list for extension extendee
-	0,   // [0:139] is the sub-list for field type_name
+	80,  // 120: wgpb.NodeOptions.nodeInternalUrl:type_name -> wgpb.ConfigurationVariable
+	80,  // 121: wgpb.NodeOptions.defaultHttpProxyUrl:type_name -> wgpb.ConfigurationVariable
+	74,  // 122: wgpb.NodeOptions.openTelemetry:type_name -> wgpb.TelemetryOptions
+	72,  // 123: wgpb.NodeOptions.prometheus:type_name -> wgpb.PrometheusOptions
+	80,  // 124: wgpb.TelemetryOptions.enabled:type_name -> wgpb.ConfigurationVariable
+	80,  // 125: wgpb.TelemetryOptions.exporterHttpEndpoint:type_name -> wgpb.ConfigurationVariable
+	80,  // 126: wgpb.TelemetryOptions.sampler:type_name -> wgpb.ConfigurationVariable
+	80,  // 127: wgpb.TelemetryOptions.authToken:type_name -> wgpb.ConfigurationVariable
+	80,  // 128: wgpb.ServerLogging.level:type_name -> wgpb.ConfigurationVariable
+	80,  // 129: wgpb.ServerOptions.serverUrl:type_name -> wgpb.ConfigurationVariable
+	69,  // 130: wgpb.ServerOptions.listen:type_name -> wgpb.ListenerOptions
+	75,  // 131: wgpb.ServerOptions.logger:type_name -> wgpb.ServerLogging
+	78,  // 132: wgpb.WebhookConfiguration.verifier:type_name -> wgpb.WebhookVerifier
+	14,  // 133: wgpb.WebhookVerifier.kind:type_name -> wgpb.WebhookVerifierKind
+	80,  // 134: wgpb.WebhookVerifier.secret:type_name -> wgpb.ConfigurationVariable
+	80,  // 135: wgpb.CorsConfiguration.allowedOrigins:type_name -> wgpb.ConfigurationVariable
+	15,  // 136: wgpb.ConfigurationVariable.kind:type_name -> wgpb.ConfigurationVariableKind
+	82,  // 137: wgpb.BuildInfo.sdk:type_name -> wgpb.BuildInfoVersion
+	82,  // 138: wgpb.BuildInfo.wunderctl:type_name -> wgpb.BuildInfoVersion
+	82,  // 139: wgpb.BuildInfo.node:type_name -> wgpb.BuildInfoVersion
+	83,  // 140: wgpb.BuildInfo.os:type_name -> wgpb.BuildInfoOS
+	84,  // 141: wgpb.BuildInfo.stats:type_name -> wgpb.BuildInfoStats
+	57,  // 142: wgpb.FetchConfiguration.HeaderEntry.value:type_name -> wgpb.HTTPHeader
+	65,  // 143: wgpb.S3UploadConfiguration.UploadProfilesEntry.value:type_name -> wgpb.S3UploadProfile
+	144, // [144:144] is the sub-list for method output_type
+	144, // [144:144] is the sub-list for method input_type
+	144, // [144:144] is the sub-list for extension type_name
+	144, // [144:144] is the sub-list for extension extendee
+	0,   // [0:144] is the sub-list for field type_name
 }
 
 func init() { file_wundernode_config_proto_init() }
@@ -7740,7 +7846,7 @@ func file_wundernode_config_proto_init() {
 			}
 		}
 		file_wundernode_config_proto_msgTypes[58].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ServerLogging); i {
+			switch v := v.(*TelemetryOptions); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -7752,7 +7858,7 @@ func file_wundernode_config_proto_init() {
 			}
 		}
 		file_wundernode_config_proto_msgTypes[59].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ServerOptions); i {
+			switch v := v.(*ServerLogging); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -7764,7 +7870,7 @@ func file_wundernode_config_proto_init() {
 			}
 		}
 		file_wundernode_config_proto_msgTypes[60].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*WebhookConfiguration); i {
+			switch v := v.(*ServerOptions); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -7776,7 +7882,7 @@ func file_wundernode_config_proto_init() {
 			}
 		}
 		file_wundernode_config_proto_msgTypes[61].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*WebhookVerifier); i {
+			switch v := v.(*WebhookConfiguration); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -7788,7 +7894,7 @@ func file_wundernode_config_proto_init() {
 			}
 		}
 		file_wundernode_config_proto_msgTypes[62].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*CorsConfiguration); i {
+			switch v := v.(*WebhookVerifier); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -7800,7 +7906,7 @@ func file_wundernode_config_proto_init() {
 			}
 		}
 		file_wundernode_config_proto_msgTypes[63].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ConfigurationVariable); i {
+			switch v := v.(*CorsConfiguration); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -7812,7 +7918,7 @@ func file_wundernode_config_proto_init() {
 			}
 		}
 		file_wundernode_config_proto_msgTypes[64].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*BuildInfo); i {
+			switch v := v.(*ConfigurationVariable); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -7824,7 +7930,7 @@ func file_wundernode_config_proto_init() {
 			}
 		}
 		file_wundernode_config_proto_msgTypes[65].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*BuildInfoVersion); i {
+			switch v := v.(*BuildInfo); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -7836,7 +7942,7 @@ func file_wundernode_config_proto_init() {
 			}
 		}
 		file_wundernode_config_proto_msgTypes[66].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*BuildInfoOS); i {
+			switch v := v.(*BuildInfoVersion); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -7848,6 +7954,18 @@ func file_wundernode_config_proto_init() {
 			}
 		}
 		file_wundernode_config_proto_msgTypes[67].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*BuildInfoOS); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_wundernode_config_proto_msgTypes[68].Exporter = func(v interface{}, i int) interface{} {
 			switch v := v.(*BuildInfoStats); i {
 			case 0:
 				return &v.state
@@ -7870,7 +7988,7 @@ func file_wundernode_config_proto_init() {
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
 			RawDescriptor: file_wundernode_config_proto_rawDesc,
 			NumEnums:      16,
-			NumMessages:   70,
+			NumMessages:   71,
 			NumExtensions: 0,
 			NumServices:   0,
 		},
